Changes to tools/hisv/hisv_events_demo_app.c
No ticket.

Changes to the pattern filter in tools/hisv/hisv_events_demo_app.c
to allow for the reading/printing of minor and informational
hardware events in addition to major and critical events.  This
change is prompted by changes to OpenHPI from version 2.12.0 to
version 2.13.3 where hotswap events on HP c-Class are now
published with the correct severity of SAHPI_INFORMATIONAL,
instead of SAHPI_CRITICAL.

Signed-off-by: Michael Bishop <michael.bishop@hp.com>