More changes to tools/hisv/hisv_events_demo_app.c
No ticket.

This change adds the ability for the tool hisv_events_demo
to decode sensor events that are of the category SAHPI_EC_SEVERITY
in addition to sensors that are of category SAHPI_EC_THRESHOLD.
The need for this change was prompted by testing sensor
events on HP Proliant rack mount hardware - but this change should
properly decode sensor events for other hardware architectures
that use OpenHPI as the source of hardware events.

Signed-off-by: Michael Bishop <michael.bishop@hp.com>