Check for null arch env variable before using in HPL library.
Ticket #435

This changes fixes a seg fault issue in the CLI when using the
hpl library and not having the OPENSAF_TARGET_SYSTEM_ARCH set
in the environment when the CLI is invoked.  This check should
have been part of changeset #435 - which is supporting HP
Proliant rack mount servers with a higher HPL MDS timeout
value than the default.

Signed-off-by: Michael Bishop <michael.bishop@hp.com>