LEAP: set line buffer mode on stdout
In context of ticket #290, the macro m_NCS_CONS_PRINTF was removed and
replaced with printf(). Since printf is be default block buffered in terminal
detached streams, the behaviour of some programs was changed.

In order to get back the old behaviour, change buffering type for the stdout
stream to line buffered. Otherwise printf output will be block buffered and
not immidiately printed to file.

This is done in a central place in LEAP which is called very early in all
daemon processes.

TODO: to be removed in OpenSAF 4.0