Removed code in hsm.c that was deleting SAHPI_ENT_RACK from entity-paths
Ticket #500

This fix removes code that was previously put in hsm.c to delete the
SAHPI_ENT_RACK entry from entity-paths that are returned by hsm.c.
Some time in the past, HP modified hsm.c to remove SAHPI_ENT_RACK
from entity-paths being returned by hsm.c. This work-a-round is no
longer needed, and, in any case, is not the proper way to deal with
this issue. The code that is deleting this part of the entity path
should be removed - and the OpenSAF user should take care to set up
their openhpi.conf file so that only correct and legal entity-paths
are used within OpenSAF.

Signed-off-by: Michael Bishop <michael.bishop@hp.com>