Disable avm and instead let avd use fm services
AVD gets the role from RDA and initialize its service from that. It also
installs an RDA callback that it uses to fail over the controller.

The AVM interface is bypassed by calling the FMA API directly. The FMA callback
is used to post a message to AVD.

HPI entity paths are built from nodeid with hard coded values for
SAHPI_ENT_PHYSICAL_SLOT & SAHPI_ENT_ADVANCEDTCA_CHASSIS. Which is fine since
AVM/HISv is not used anyway.

The AVM thread within scap is not started.