scripts: exit status when program is not running should still be 0 (#1858)
Per LSB 20.2. Init Script Actions:

...the init script shall return an exit status of zero if the action was
successful. In addition to straightforward success, the following
situations are also to be considered successful:

  * running stop on a service already stopped or not running

In this case 'killproc' on a crashed opensaf daemon would exit 7
(7 program is not running), and bypass lockfile cleanup. This
was causing problems at AMF instantiate if the lockfile was still
present, preventing the daemon to restart properly.