avsv: move log from mds callback to main (#2706)
Move a log in avnd from MDS callback context to the main thread.
Having the log in the MDS callback caused confusion whether
the amfd process actually had crashed when it was in fact killed
by an action of amfnd. Such as rebooting the local node due to
some failure.