glsv: Replace faulty lock macro with pthread_mutex (#2618)
This bug was originally detected in LEAP, and it is triggered when the
application calls saXxxInitialize() from two threads concurrently (either for
the same SAF service or for two different ones). The problem was that there is a
flag indicating if the mutex has been created, but this flag must itself be
protected by a mutex. Hence, this construct is not thread-safe. When two threads
enter this code, they both read the flag saying the mutex needs to be created,
so they both create the mutex and enter the critical section at the same time.