tools/safimm: clear parameter bindings after prepared sql statement execution (#2911)
After execution of a sql prepared statement with variable number of parameters, clearing parameter bindings is not done, and it causes that if some parameters in a next execution of the same statement are not bound with new values/pointers, old values/pointers bindings are used for the new execution of the statement.
The new patch clears sql prepared statement bindings after reseting sql statement with variable number of parameters.