safimm: Add variant of function ccbutil_ccbAddCreateOperation (#3001)
A variant ccbutil_ccbAddCreateOperation_2(....) is added. This variant has
the same function as the original function with the addition that it takes
the parameter objectName and adds the value to the CcbUtilOperationData struct
in the same way as the Delete and Modify operation.

The immutil in the samples directory is updated to be the same as immutil in
tools