uml: Fix race condition when creating the uml_switch socket (#3025)
Sometimes when starting UML, some of the nodes were unable to connect
to the uml_switch socket. This had the effect that the eth0 network
interface was missing in those nodes, and thus OpenSAF could not be
started.

This fix makes the handling of the uml_switch socket more robust,
by making sure the socket has been created before starting the nodes.