immsv: IMMND Nack's to NID at controlled failure (#3039)
IMMND sync client that detects a problem during sync does not just
exit/abort/osafassert because the NID will then only restart the IMMND
for a new sync after the NID times out on its wait on the IMMND start.
Instead the IMMND will send a NACK to the NID before terminating.