avsv: harden avnd_evt_avd_info_su_si_assign_evh (#2921)
In some error escalation scenario that ends in node failover,
avd send a SUSI delete message for a certain SI. For unknown
reasons this SI does not exist in the database and avnd asserts.
Since avnd anyway is about to send a fake response to SUSI delete,
the assert makes no sense. Better just check the db and log a warning
in case the SI does not exist.