avsv: ignore comps in FAILED state at shutdown (#2965)
Previously if a single component had a presence state of INST-FAILED or
TERM-FAILED, amfnd would call opensaf_reboot during system shutdown.
Consequently if an issued cleanup command failed, opensaf_reboot would be
called. This can cause escalations to failfast reboot although a user just
tried to stop opensaf or even halt the system. But a TERM-FAILED state is bad
because AMF has lost control over the execution environment and a reboot is
needed. This is to get rid of any orphaned component processes that its
corresponding cleanup command could handle.

The calls to opensaf_reboot are removed for comp events during shutdown. To
ensure that opensaf is not restarted with a comp in TERM-FAILED pres state,
a new "failed state file" is introduced. When a component enters TERM-FAILED
pres state, the file is created in a location cleared by the boot process
(typically /var/run). When opensaf is started and the state file exist, the
start will wail. Either a manual cleanup is needed (incl the state file) or
a reboot performed before opensaf can be started again.