avsv: keep IMM jobs in queue in noactive state (#3028)
Calling e.g. saImmOiRtObjectUpdate_2() in the NOACTIVE state is a bad idea after
the #2446 change. The call will return TIMEOUT, next attempt BADHANDLE and then
the thread for re-initialization fails with ERR-EXIST on saImmOiImplementerSet()
and the amfd process exits.

The patch keeps jobs in the queue in the no-active state. After that, jobs are
flushed to IMM by the normal background job function.