immsv: implementerSet optimistically sets client node flags (#3072)
Optimistically turn on cl_node->isApplier or cl-node->isPbe when
applier/pbe is trying to set its implementer-name. These flags are
set to 'on' when relevant before the implementer-set request is
is sent by the imma library to the server. If error is returned,
the flags are reverted to zero. If OK is returned then the flags
are already correctly set. This will allow the mds thread to receive
relevant applier or pbe callbacks before the main thread has received
the OK reply on the implementerSet request.