immsv: saImmOiImplementerSet returns INVALID_PARAM on interface violation (#3086)
The reason for the obeserved crash was (1) an erronesous try by the OI
to set implementer on a handle where implementer had already been set by
that process on that handle; (2) this user error was handled incorrectly
by the IMMA library by returning ERR_EXIST (which is an error that does not
signify interface violation and encourages retry, see ticket analysis).
Finally (3), the reason for the actual crash was that the IMMA code also
cleared flags set optimistically by the call. The problem was that due to this
call being both redundant and incorrect, the optimistically set flags where
actually already set correctly, by the prior correct call.  When a callback
arrived on the previously correctly set implementer, there was an assert due
to the now incorrect flag state.