opensaf: Fix build problems with GCC 4.8.0 (#3098)
Fixed two type of compiler warnings:

hj_edp.c: In function 'ncs_edu_ncs_key_test_fmat_fnc':
hj_edp.c:1592:4: error: typedef 'LCL_TEST_JUMP_OFFSET' locally defined but not used [-Werror=unused-local-typedefs]
  } LCL_TEST_JUMP_OFFSET;

This one is simply complaining that we have a typedef locally within a function, and it is never used. Easy to fix, simply remove the typedef.

ncs_mda.c: In function 'mds_fixed_vdest_to_inst_name':
ncs_mda.c:349:26: error: argument to 'sizeof' in 'memset' call is the same expression as the destination; did you mean to dereference it? [-Werror=sizeof-pointer-memaccess]
  memset(o_name, 0, sizeof(o_name));

This one complains that the last parameter to memset (the size) most likely is wrong. Indeed, it is a bug; we are taking the size of the pointer rather than the size of the object it is pointing to. Should be:

  memset(o_name, 0, sizeof(*o_name));