opensaftools: Fix diffstat generated by review script [#3]
When specifying a range of changesets to the OpenSAF review script, the wrong
diffstat was shown if the parent to the first changeset in the range does not
have the changeset number immediately before it (i.e. because that changeset
is on a different branch). Now the correct parent is selected.