avsv: allow two SUs on same node in nwayactive (#3066)
* * *
avsv: fix nwayactive to support node lock with more than 2 SUs (#3066)
* * *
avsv/avd: properly count number of sent asgn msgs, node unlock (#2066)

This change reverts and improves the fix introduced in ticket 2660. In 2660
IMM was responded only when the SG is stable. That was needed because number of
SUSI asgn messages was not properly counted.

This patch introduce proper counting of messages sent, therefore the admin
response can be done when the counter reaches zero.