amfnd: remove premature nid response (#45)
In the scap split (842:57daaf68acd5) a response to NID was added before entering
the main loop in amfnd. The problem with that is that it leaves amfnd
unprotected until it has started the AMF watchdog component.

By removing this NID ack, the amfnd startup gets supervised by NID and any crash
in this phase will result in NID taking action as configured.