osaf: Add time supervision of opensaf_reboot [#437]
Add a time supervision of the library function opensaf_reboot() as well as the
shell script opensaf_reboot. If the reboot has not happened before the timeout,
the OS is rebooted hard using the SysRq trigger /proc/sysrq-trigger. This makes
it possible to reboot the node also when the system is in a very bad state, for
example when fork() fails because the system is out of resources (no free
memory, process table full etc.).  It also handles the case when the ordinary
reboot command hangs trying to sync the file system, for example due to a disk
or NFS problem.

Note that only amfwd and amfnd should reboot the local node. Other services
should simply call _Exit() and let amfnd handle the reboot.