IMM: Correction to ccb handling in immnd_fevs_local_checks() [#535]
Patch only for 4.3 and default(4.4).
NOT for 4.2 branch

The logic in immnd_fevs_local_checks() in immnd_evt.c is corrected
for CCB related messages to use immNotPbeWritable and not immNotWritable()
as the precheck. The former is unnecessarily strict since the intention
is to allow ongoing ccbs a period of grace to complete before sync starts,
but not allow new ccb-id's to be generated. The logic here also did not
recognize the special encoding of TRY_AGAIN needed towards the library for
CcbApply.