IMM: Ignore spurious/redundant ccb-apply from client [#535]
A spurious and redundant ccb-apply generated by a faulty code elsewhere shall
be ignored by the immnd server.
Bugs in the imma library or new bugs introduced in the immnd server in the
future should never be able to cause such a serious error as an inconsistent
commit. The problem here was that the arrival of the second apply reaching the
ccb already in critical at some processor(s) caused the ccb to get *aborted*
in imm-ram while being in critical. That should never be allowed to happen.
A ccb in critical is waiting on the commit decision from PBE and PBE alone.
Patch also fixes a potential source of redundant apply in imma_om_api.c
potentially caused by mutithreaded applications (incorrect usage of handle).