ntf/test: use stat on ntfsv_test_classes.xml before immcfg [#575]
Do not use a negative response from immcfg as an implicit method
to find the ntfsv_test_classes.xml file. A negative command
response is interpreted by the test framework as if the whole
test case failed. Instead, use stat() to check if the file is located
in /tmp directory, and if not there, search for it and put it in /tmp.
Then proceed with immcfg.