amf: change saAmfResponse to be synchronous in TerminateCallback context [#570]
If a component crash or exit in context of the terminate callback, AMF will not
use the "ava down" event to trigger cleanup and finish component termination.
Instead the CallbackTimeout is awaited which can be very long.

This is a big problem if this happens during an upgrade, it will cause the
upgrade to fail potentially leading to system restore.

By making the saAmfResponse call synchronous for the TerminateCallback, the
"ava down" event can be used to trigger cleanup of the component.