IMM: IMMD file verification made upgrade safe [#596]
The protocol for communicating IMMND file configuration to the active IMMD
and for checkpointing the same from active to standby IMMD has been elaborated
and made safer. In particular, an upgrade from pre 4.4. OpenSAF to OpenSAF 4.4
is now possible.

The file configuration consists of:
(1) The directory for peristent storage,
(2) The xml file to load from if PBE is not enabled or if no PBE file exists,
(3) The pbe file base-name to load from if PBE is enabled.

The IMMNDs also communicate epoch and PBE enabled/disabled state to the IMMD
and the active IMMD checkpoints this to the standby.