IMM: 2PBE patch-5 (Fix loading retry problem) [#21]
If after loadin-arbitration is done, loading is started but fails (for example
due to a corrupt sqlite file) then the IMMNDs are restarted but not IMMDs.
The already collected loading arbitration info in the active IMMD is not
cleared and in the next loading attempt the loading arbitration will only wait
for the stats from one of the SCs, then find it already has stats from both
SC's but actually one of them will be stats from the previous load (!).
This can result in incorrect arbitration. The file for which loading fails
will have been moved to imm.db.xxxxx.failed and thus not used for arbitration.
The fallback file is typically much older, but that may be masked by the
old preload stats for that SC. The preload stats need to be cleared here
when and if loading is restarted.