IMM: Support for 'immadm -a (--admin-owner) <admo-name>' [#613]
Allow user to request immadm to set admin-owner to a specific admin owner name.
By default, a temporary admin-owner name is allocated by the immadm command.
The '-a' argument allows the user to choose the explicit admin-owner name that
shall be set for the object on which the administrative operation is to be
invoked on. This is typically only needed if the object already has admin
owner set, more or less permanently. Because of this, no admin-owner-release
will be done by immadm when explicit admin-owner name is used.