IMM: 2PBE patch-1 (loading) [#21]
This patch contains the 2PBE loading mechanism, needed to support 2PBE.
The IMMD's will detect 2PBE loading by the IMMSV_2PBE_PEER_SC_MAX_WAIT
environment variable being set in the immd.conf file. The active IMMD
will order each SC IMMND to execute a "preload" probing the SC local
filesystem for the file state that would be loaded to the cluster if that
SC IMMND was chosen as coord. The two SC IMMNDs sends the preload stats to
the active IMMD.

The active IMMD will wait for the IMMNDs at *both* SCs to complete the
preload task and then determine which SC has the apparently latest file
state. The IMMND at that SC will then be chosen as IMMND coord.

Actual loading then proceeds in the same way as for regular 1PBE.
The IMMSV_2PBE_PEER_SC_MAX_WAIT is by default 30 seconds. This value
should be high enough to make it extremely unlikely that the active IMMD
is forced choose loader when only a single SC IMMND has joined.
If that happens, then the risk is that the cluster restart will be
done *not* using the latest persistent imm state, effectively rewinding
the imm state. (Note: the same thing will happen with regular 1PBE based
on a shared filesystem (DRBD) if the one SC fails to come up in time to
join the DRBD sync protocol. The corresponding DRBD timeout is on the order
of 20 seconds).

When loading has completed, additional 2PBE functiolaity will start two PBEs,
one at each SC. That functionality is delivered in subsequent patches.