IMM: 2PBE patch-3 (1safe2pbe) [#21]
This patch contains the 2PBE 1safe2Pbe mechanism. This mechanism
allows an OpenSAF cluster to open up for persistent writes using
only one of the two PBEs - temporarily.

This is only intended to be used as an emergency action when one SC
is long term unavailable, e.g. hardware problems. As soon as the other SC
returns, the IMM has to re-enter normal 2-safe 2PBE and reject persistent
writes until the slave PBE has synced (regenerated its sqlite file) and
rejoined the cluster.

The 1safe2PBE state is entered by the administrative opeation:

  immadm -o 1 -a safImmService -p opensafImmNostdFlags:SA_UINT32_T:8 \
     opensafImm=opensafImm,safApp=safImmService

It is exited either automatically by a rejoined SC or by an
explicit administrative opertion:

  immadm -o 2 -a safImmService -p opensafImmNostdFlags:SA_UINT32_T:8 \
     opensafImm=opensafImm,safApp=safImmService