IMM: 2PBE patch-4 (Fix for si-swap problem) [#21]
With the 2PBE patches applied and 2PBE configured, if si_swap:

     immadm -o 7 safSi=SC-2N,safApp=OpenSAF

is attempted twice, then the second time will cause the new active SC to reboot.
This was caused by the cb->is_loading variable being initialized to true at
both active and standby SC, when it should only have been set to true in
the active. When true also in the standby it is not set to false by loading
completed. When the standby becomes active, it will not mbcp fevs messages
to new standby, despite that loading was done a long time ago. Next si swap
causes the new new active to start sending fevs messages way below the
expected fevs number. This causes the new new standby to crash.