IMM: 2PBE patch-7 (Fix two problems in immloader) [#21]
The 2PBE patch stack introduced two problems in immloader.

The first problem was that default loading without PBE enabled no longer
was working. Initialization of an STL string parameter in main() for
immloader was not done correctly. This caused immloader to crash whenever
started with pbe not enabled.

The second problem was that the introduced creation of an IMM runtime
class dedicated for supporting IMM PBE, done by immloader, did not
screen for the preloader case when preloading was done from xml-file.
The immloader tried to create this new special runtime class already
in preloading, which of course failed.

Both of these problems are fixed in this patch.