amfd: fix admin op logging [#85]
If immadm or amf-adm is used when invoking an AMF admin op, error messages end
up on stderr. But since SMF is not adopted to use admin op error strings, no
logging at all will be done if an admin op fails during campaign execution.

This patch logs the error string to saflog before respondning to the admin
operation.