IMMND replies BAD_HANDLE if client node is missing for non fevs request [#637]
For non fevs requests, if the local IMMND can not find the client node
corresponding to the imm-handle sent by the client, then the local IMMND
replies directly with SA_AIS_ERR_BAD_HANDLE, instead of just dropping
the request and causing client to timeout. The IMMND also prints a warning
log message with the handle value included.

For fevs requests, this short circuited behavior is only possible  if the
client is missing when the originating request arrives at the local IMMND
in the fevs_forward function. If the client exists at that time, but is
gone by the time the fevs request arrives back (from the IMMD) at the
originating IMMND, then no reply can be sent. The originating IMMND no
longer has the reply information on the stack. It expects to find it in
the client node on the heap.