base: Add osaf_poll.h and osaf_time.h APIs [#580]
Add new utility and convenience APIs, declared in osaf_poll.h and osaf_time.h,
respectively.

osaf_poll.h contains utility functions that work in a similar way as the Linux
function poll() and ppoll(), except that they handle errors themselves instead
of returning -1. The errno value EINTR is handled with a loop, and the functions
keep track of time so that the time-out will happen at the same time no matter
if the functions were interrupted by a signal or not. Other errno values will
cause the process to be aborted, since they indicate the presence of a software
bug and the program cannot continue in a safe way.

osaf_time.h contains utility functions for reading the time and sleeping, that
work in a similar way as the Linux functions clock_gettime() and nanosleep(),
except that they abort the process instead of returning -1. There are also
utilty functions for manipulating "struct timespec" times. struct timespec is
used for representing a time in many POSIX functions, and the reason for having
a structure is that a 64-bit integer is not sufficient for representing the full
time.