IMM: Discard node for self changed from osafassert to LOG_ER & exit [#60]
If an IMMND receives a discard-node message and the node to be discarded
is the node the IMMND is executing on, then this is a clear indication of
cluster network partitioning, or "split brain". The action taken by the IMMND
was to osafassert if this happened. But such an assert generates a coredump
resulting in unnecesary tickets on and troubleshooting off the IMM.

With this patch, the error is instead log'ed to the syslog and then IMMND exits.
It should then be clear why the IMMND restarts and that it was not due to an
error in the IMMND.