java: Fix compiler warnings and treat warnings as errors [#649]
Makefile.common was not included by the makefile for the Java bindings, so
warnings were not treated as errors. It is now included, and all compiler
warnings in the Java bindings have been fixed. Two of these warnings were
actual bugs: the wrong function was called (U_printSaClusterNotification
instead of U_printSaClusterNotification_4, and U_printSaClusterNode instead
of U_printSaClusterNode_4).