clm: fill saClmNodeAddress and saClmNodeAddressFamily with ip info [#228] Ticket #617 enhanced the MDS NODE_UP event to additionally deliver the ipaddress and address family information. Currently, these attributes by default carry SA_CLM_AF_INET(i.e. AF_INET) as the address family with values as zero. There is already code to carry these attributes in the track callback and node_get/node_get_async.
This patch obtains the ip information from MDS and populates these values
in the track callback and nodeGet APIs, i.e. When TCP is the mode of transport.
When TIPC is the mode of transport then the address length will be set to
zero and the family will be set to SA_CLM_AF_INET(i.e. AF_INET).
This is done for backward compatibility with pre 4.4 libraries
and because the CLM spec does not support any value other than
SA_CLM_AF_INET(i.e. AF_INET) and SA_CLM_AF_INET6 (i.e. AF_INET6)