amfnd: guard failed state file removal [#741]
When SGAutoRepair is enabled and performed the following syslog can be observed:

Jan 24 07:41:57 Vostro osafamfnd[17227]: ER cannot unlink failed state
    file /var/run/opensaf/amf_failed_state: No such file or directory

The same happens when repairing an SU in INSTANTIATION-FAILED state.

The reason for this that the failed-state file removal is done in all cases.

The patch protects the failed-state file removal so that it is only done in the
case of the SU being in presence state TERMINATION-FAILED (which is the only
state the file can exist in).