amfnd: escalate comp errors in quiesced HA state [#601]
When a node or SU is locked, AMF assigns the QUIESCED HA state to ACTIVE
sa-aware components. This state is intended to transfer state to a peer
component assigned STANDBY. If the ACTIVE component fails in the QUIESCED
state, AMF will restart it and reassign it QUIESCED. This is an invalid state
transition according to picture 3 in the B.04 AMF spec and makes no sense since
the state is anyway gone (because the component failed).

The problem was likely reintroduced when fixing #3083, before that special
handling for comp errors in QUIESCED state was there. However the solution had
other problems and was removed. So in the current code there is no special
handling for errors in QUIESCED state which results in re-assignment of the
QUIESCED HA state after component restart.

This patch changes the error escalation logic so that instead of component
restart, component failover is performed as recovery action.