amfd: create comp if not found during decoding [#761] Standby controller reboots during execution of an installation campaign. The following be seen in syslog: Jan 30 08:19:53 err sc2 osafamfd[15184]: avd_ckpt_dec.c:3163: avsv_decode_warm_sync_rsp: Assertion '0' failed. When the operational state of an added component gets ENABLED, this is checkpointed to the standby AMFD. If at this point the component does not exist in the standby database, the count of async updates is not updated. Later when the mbcsv warm sync happens, the mismatch in async updates count is detected and the assert happens. This can happen when the applier callbacks has not yet happened on the standby AMFD. This is a pure timing problem.
We fix this by creating the component in the database if it does exist just for the purpose of storing the runtime attribute (op state). Later it will be filled in with configuration attributes when the applier callback eventually is called. The async counters stay in sync and the mismatch is not detected at warm sync.