amfnd: prepend prefix to clc-cli cmd if exist [#773]
Component fails to instantiate, typical syslog:
Feb  7 16:31:25 Vostro osafamfnd[19923]: ncs_os_process_execute_timed:
    execvp '/amf_demo_script' failed - No such file or directory

If the saAmfCtRelPathInstantiateCmd attribute starts with a "/" the path prefix
from the possibly corresponding SaAmfNodeSwBundle instance is ignored thus
creating an invalid path.

We change amfnd to always prepend the command with the path prefix if it exist.