IMM: 2PBE slave uses main/applier thread for PRT data flow [#623]
The main cause for the unclean cutoff was that the 2PBE slave used the
runtime thread for dataflow for PRTO-create, PRTO-delete and PRTA-update.

The termination of PBE is triggered by a regular CCB commit, which
arrives via the main/applier thread in the 2PBE slave. As soon as the
PBE termination CCB has comitted in imm-ram, the IMMNDs at the SCs will
order their PBEs to terminate by sending a SIGTERM to their PBE process.
The signal is caught and translated to an event on a termination descriptor
polled by the main/applier thread.

Because the 2PBE-slave did essentially all processing of PRTO ops
in the runtime thread, they would esily be interrupted and terminated
from the main/applier thread, before completing their current PRT job
in the runtime thread.

Such interruption/termination of the process, by a searate thread is
very timing sensitive. This sometimes caused a discrepancy in termination
between the primary PBE and the slave PBE.

By transferring all the data flow for PRTO operations to the main/applier
thread in the slave, any on-going PRTO job in the slave will complete before
the termination job is excecuted by the main-thread.

The spinlock mechanism added for 2PBE to synchronize the use of sqlite
transactions between the slave's main/applier thread and runtime thread
has been elaborated to eliminate the risk of the sqlite transaction
being comitted before the its transaction buildup has been completed.
This is needed now because the buildup of PRTO-create and PRTA-update
in the slave is executed by the runtime thread, but the sqlite commit in
the slave is handled by the applier/main thread. So the main thead has to
wait not just for the start of the transaction, but also for the completion
of the prepare/buildup.

This patch also does a bit of minor cleanup in the fuctions touched.