amfnd: report failover after comp cleanup [#598] If a component error is detected and the recovery action is COMPONENT_FAILOVER, it is possible that a standby component gets the active assignment before the erroneous component has been terminated. This can cause a split brain on application level.
The reason for this is that when the error is detected amfnd starts two
parallel activities, component cleanup and inform director. When the director
receives the information it starts the process of failing over the workload
of the erroneous component.

This patch informs the director after successful termination has been performed.