IMM: Use existing PBE mutation record to guard against concurrent class create/delete [#757]
The existing mutation record for PBE operations includes records for class-create and class-delete.
These are now used to detect any pre-existing and on-going PBE job for a class-create or class-delete.
When this is detected, any new request to delete or create the same class will get ERR_BUSY.

If PBE is not enabled, then class-create and class-delete are pure uninterrupted single jobs
in the IMMNDs and have no concurrency issues.