amfnd: fix failover of NPI SU during SU lock with fault [#820]
Problem: During lock of NPI SU, AMF does not failover assignments
if a healthy component faults.

Reason: During SU lock AMF sends quiesced assignment to NPI SU.
During quiesced assignment, when termination of one component is in progress,
one of the healthy components faults. AMF performs cleanup of this failed
component and it moves the SU from TERMINATING to UNINSTANTIATED state
without waiting for the terminate command response of the first component.
Thus SU moves to UNINSTANTIATED state when one of the components is still
in TERMINATING state.  Now when first component responds for the terminate
command, AMF sees SU in UNINSTANTIATED state and it does not take any action.

Fix: AMF should move NPI SU in UNINSTANTIATED state only when all the CSIs
are moved to Removed/Assigned state.