IMM: Special applier callbacks for PRTO deletes are re-established [#835]
Changes to the server code for immsv done in relation to adding support
for 2PBE in OpenSAF 4.4, broke the logic that generates callbacks to
special appliers for PRTO deletes. The problem exists with 0PBE, 1PBE or 2PBE.

This changeset fixes the problem so the special applier callbacks
for PRTO deletes are now sent.