avnd: send SUSI response after cleanup of comp [#822]
During component fail-over a standby component can be activated before cleanup
of the faulty component has finished effectively introducing split brain on
component level.

This happens because cleanup is not awaited before the SUSI response message is
sent to the director.

Fix this by sending the response after the cleanup has finished.