amfd: act on dependent if sponsor gets unassigned in NoRed [#832]
Problem: Due to fault in stable state in NoRed model, if  sponsor goes
unassigned then AMF is not unassigning dependent.

Reason: Due to fault like node-failover or su-failover in stable state,
if sponsor can not be fail-overed to another SU, AMF deletes the SUSI
of sponsor. At the same time AMF is updating the si_dep state of
dependent to READY_TO_UNASSIGN, but it is not taking action on the
dependent SI. In such a case a user will see the removal of assignment
of dependent directly or via expiry of tolerance timer.

Fix: This patch ensures that if sponsor is not fail-overed to any spare
SU, then it should take action on dependent SI.