avsv: include and use sirank in SUSI msg [#574]
amfnd process aborted in immutils and node rebooted by amf watchdog with the
following log:
2013-09-09 11:50:02 osafamfnd SC-2-1 err osafamfnd[5336]: saImmOmInitialize FAILED, rc = 5

When the AMF node director receives a SUSI ASGN message, it reads SI rank
from IMM. If IMM does not respond in a timely manner the amfnd process will be
aborted and the node restarted by the AMF watchdog.

By including SI rank (an int) in the SUSI ASGN msg, the read from IMM can be
skipped/done conditionally based on the amfnd MDS version.