avsv: include and use comp capability in SUSI msg [#574]
When the AMF node director receives the SUSI ASGN message, for each CSI it reads
the saAmfCSType attribute from IMM.

When the AMF node director receives the SUSI MOD message, for each CSI it reads
the saAmfCtCompCapability attribute from the comp and csi associated
SaAmfCtCsType instance.

In any of these cases if IMM does not respond in a timely manner the amfnd
process will be aborted and the node restarted.

By including component capability (an int) in the SUSI ASG msg, these reads
from IMM can be completely skipped in the node director.