amfnd: update saAmfSURestartCount & saAmfCompRestartCount during escalations [#847]
Problem: AMF does not update saAmfSURestartCount and saAmfCompRestartCount
when error escalation is going on or when timer related to error
escalations expires.


Reason: AMFND is not sending update of these attributes to AMFD. Due to this,
AMFD does not return updated value to IMM upon request.

Fix: AMFND will send updates of these attributes to AMFD.