amfd: fix imm job update [#868]
Ugly logs from amfd like:
Apr 19 19:55:18 SC-1 osafamfd[469]: ER exec: update FAILED 12

When an IMM update job is executed and for some reason the object does not exist,
the job remains in the queue. Reason is because the wrong error code is checked.

By correcting the check, the job is removed and no logging appears.