amfd: fix infinite loop when processing CLM node tracking [#877]
Logs from amfnd:
Apr 25 21:37:13.521696 osafamfnd [4547] ER AMF director heart beat timeout, generating core for amfd

clm_node_exit_validate loops through all the SUs for the passed in node, but
never actually moves to the next SU, so the while test is always true.

By moving on the the next SU in the loop, we can exit the while loop when there
are no more SUs.