amfd: cleanup SU [#713]
su_act_state is not used thus can be made a dummy (not removed because of EDU
usage). The corresponding enum AVD_SU_STATE removed. Checkpointing of
su_act_state kept for backwards compatibility but reduced to a minimum.

The use of "this->" is removed from some setters.

Some non existing function signatures removed from the header file.