log: fix errno usage for calls to strtoul [#841] On UML based on ubuntu, the following code fails consistently.         if ((val_str = getenv(LOGSV_MAX_LOGRECSIZE)) != NULL) {                 val_uint = strtoul(val_str, NULL, 0);                 if ((errno != 0) || (val_uint > UINT_MAX)) {                         LOG_ER(Illegal value for LOGSV_MAX_LOGRECSIZE -.....,                         strerror(errno), lgsConf->logMaxLogrecsize);                         lgsConf->logMaxLogrecsize_noteflag = true;
i.e. the errno is always getting set to a nonzero value.
Because of this, when user configures an environment variable, the above code fails causing
log to assign default values rather than the values from the environment variables.
As per strotul, errno = 0 has to be done before calling strotul.
The following NOTE from the man page of strtoul is for reference:
NOTES
       Since strtoul() can legitimately return 0 or ULONG_MAX (ULLONG_MAX  for
       strtoull()) on both success and failure, the calling program should set
       errno to 0 before the call, and then determine if an error occurred  by
       checking whether errno has a nonzero value after the call.
.
With this patch, when a environment variable is configured(when the default configuration object
is not configured) then LOG shall use the value specified in the environment
variable instead of setting default values.