amfnd : avoid multiple clean up of comps during shutdown phase [#890].
During shutdown phase, AMFND performs multiple cleanup of components
in a SU if one of the components moves in TERM_FAILED state.

AMFND launches clean up of all the components during shutdown phase.
If cleanup of one of the components fails then AMFND again launched
cleans up all the components of the SU as recovery. In this way mutliple
clean ups are performed for the same component.

Patch checks term_state of avnd control block before launching clean up
of components in the SU.