amfd: allow increase of saAmfSGNumPrefActiveSUs for N+M SG while UNLOCKED [#871]
ccb_completed_modify_hdlr: Attribute 'saAmfSGNumPrefActiveSUs' cannot be
modified when SG is unlocked

OpenSAF disallows changing saAmfSGNumPrefActiveSUs while the SG is UNLOCKED.
This prevents in-service capacity upgrade for N+M models, which is not
desirable.

This change allows saAmfSGNumPrefActiveSUs to be increased for N+M models,
as long as there are instantiated spare SUs equal to the amount of the increase.
These instantiated spare SUs will then be assigned ACTIVE, and their
corresponding CSIs will be assigned to the STANDBY.