clm: pass client_id for deletion; post to mbox only when svc_id is filled [#888] Valgrind reported the following: ==9425== Invalid read of size 4 ==9425== at 0x40AAE2: clms_client_delete_by_mds_dest (clms_evt.c:167) ==9425== by 0x40AC50: proc_clma_updn_mds_msg (clms_evt.c:621) ==9425== by 0x40BB8F: clms_process_mbx (clms_evt.c:1307) ==9425== by 0x403FDD: main (clms_main.c:499) ==9425== Address 0x67ba190 is 32 bytes inside a block of size 64 free'd ==9425== at 0x4C2BDEC: free (in /usr/lib/valgrind /vgpreload_memcheck-amd64-linux.so) ==9425== by 0x40AA22: clms_client_delete (clms_evt.c:205) ==9425== by 0x40AAE1: clms_client_delete_by_mds_dest (clms_evt.c:164) ==9425== by 0x40AC50: proc_clma_updn_mds_msg (clms_evt.c:621) ==9425== by 0x40BB8F: clms_process_mbx (clms_evt.c:1307) ==9425== by 0x403FDD: main (clms_main.c:499)
This warning from valgrind is shown in the context of the below lines:
rc = clms_client_delete(client->client_id);

 /* Delete this client data from the clmresp tracking
list */ rc =
clms_client_del_trackresp(client->client_id);

It can be seen that clms_client_del_trackresp was accessing a memory that
would get freed in some flows by the previous line.
The patch avoids accessing the freed memory and converts the local variable
client_id to hostorder before passing to clms_client_del_trackresp().
The second part of the patch adds a check in clms_mds.c such that
posting to mailbox is done only after explicitly checking whether
the service id is for CLMA or AMFND.
There is a subscription for CLMNA also, but we are not doing anything today
for CLMNA down. So, in the case when CLMNA down comes, the event-type
field is not getting set when posting to the mail box. This results
in a situation where the mailbox processing invokes proc_initialize()
because the event-type gets set to 0 which is the default value.