amfnd : fix SU restart escalation [#885].
During SU Restart recovery all components are not restarted in
NPI SU.

In case of npi SU restart recovery, this problem appears because the condition
of changing su presence state to INSTANTIATED is not sufficient.
If the component of the last csi in csi_list has been restarted first due
to componentRestart (before suRestart), AMFND can not find any next csi
and changes the SU presence state to INSTANTIATED, this will miss the
restart for the rest of components.

The fix uses UNASSIGNED csi to avoid duplication of restarting the same
component many times during su restart. Only component linked to UNASSIGNED
csi is restarted. Therefore, the condition of changing su presence state to
INSTANTIATED should be all csis are ASSIGNED.