imm: Add upgrade support for saImmOmValidate [#798]
Symptoms are during upgrade to OpenSAF 4.5:
May 16 12:01:03 SC-2 local0.err osafimmnd[427]: ER Can not sync Ccb that is active

The cause of the problem (expected actually) was that upgrade support had not yet
been added to default (4.5). Addition of the mechanism for supporting upgrade of
immsv to OpensAF 4.5 is tracked by ticket #842.

Solution is first to get the implementation of [#842] pushed, then to get this
changeset pushed. This changeset uses the mechanism provided in [#842] to (a)
onlu allow saImmOmCcbValidate if protocol45 is allowed: and (b) revert to the
pre 4.5 protocol for syncing ccb-states if protocol45 is not allowed.
Note that saImmOmCcbAbort is not a problem because it does not depend on any
new message types.