amfnd: fix process of proxy proxied cleanup on OpenSAF shutdown [#629]
Durinhg opensaf shutdown node reboots, if it hosts proxy and proxied components.

On OpenSAF shutdown, amfnd does not wait for the proxied components of a proxy to
terminate before terminating the proxy.  It terminates the proxy via CLC-CLI
cleanup, and the proxied component via the
SaAmfProxiedComponentCleanupCallbackT in the proxy.  If the proxied component
takes some time to terminate, amfnd doesn't wait and terminates the proxy.
amfnd never gets the response from SaAmfProxiedComponentCleanupCallbackT
(because the proxy has been terminated), and thinks that the termination of the
proxied component timed out.

amfnd needs to delay terminating the proxy until all the proxy's proxied
components have terminated.