base: Corrected handling of raise_obj, rmv_obj file descriptors of Selection object [#928]
This patch addresses the corrections suggested by the ticket
#928, in handling raise_obj, rmv_obj file descriptors of selection object.

1. Whenever the file descriptors raise_obj and/or rmv_obj are closed, the
respective file descriptor(s) in the data structure are set to -1 indicating
that the file descriptor(s) are no longer valid.

2. The function ncs_sel_obj_rmv_ind() checks rmv_obj file descriptor, and if
set to -1, return immediately without trying to operate on the file
descriptor.

3. The function ncs_sel_obj_ind() checks raise_obj file descriptor, and if
set to -1, return immediately without trying to operate on the file
descriptor.

4. If still recv() fails due to any other reason than EAGAIN or EINTR, does
osaf_abort() to generate a core dump.