avd: default saAmfSGNumPrefInserviceSUs to unlimited [#278]
When an SG is extended with a new SU, it will not be instantiated or
assigned unless saAmfSGNumPrefInserviceSUs is configured accordingly.

Reason is saAmfSGNumPrefInserviceSUs is set to number of SUs when the SG is
created and is not recalculated as SUs are added. Forcing the user to configure
saAmfSGNumPrefInserviceSUs instead of leaving it unconfigured and relying on
the default value SU count as specified.

This patch sets saAmfSGNumPrefInserviceSUs (when not specifically configured)
to unlimited. This mean when an SU is added, nothing needs to be done. It will
just be instantiated since number of instantiated SUs is always lower than
unlimited.