amfd : standby AMFD should process ccb completed cbk for delete [#947]
Standby AMFD crashes when it executes CCB apply for delete.

This got missed in #305. In CCB delete AMFD memorises the object to be deleted.
When applier comes, it uses this stored pointer to delete the object.
In #305, standby AMFD was prevented to execute any completed callback.

This patch allows CCB completed callback for deletion at standby AMFD.