IMM: PBE allows config changes on opensafImm=opensafImm,safApp=safImmService [#934]
The OpenSAF IMM service object 'opensafImm=opensafImm,safApp=safImmService'
currently has two writable config attributes:

   - opensafImmSyncBatchSize
   - longDnsAllowed

Changes to these attributes required the PBE to be disabled. This because the PBE
is both regular-OI and PBE-OI for this special object. The PBE used to reject all
such requests. This ticket changes the PBE to accept modifications to this config
object. This would result in the PBE getting a second (duplicate) callback, the
first one as regular OI and the second one as PBE, resulting in the derailing and
abort of the CCB. So this changeset also fixes so that modifications to the special
OpenSAF IMM service object will only generate one modify callback to the PBE. The
regular OI callback is discarded for this special case.

The ticket also mentions that validation should be done. But there is no validation
of substance to perform for the above two config attributes.
The 'opensafImmSyncBatchSize' can be set to whatever value the deployer thinks is
appropriate. The 'longDnsAllowed' attribute is managed by logic that only cares if
the value is 0/zero or not, corresponding to longDNs disallowed or allowed. So the
PBE simply accepts whatever modification is proposed for these attributes.

Creates or deletes of instances of the class OpensafImm are still rejected by the
PBE (not changed by this changeset) and so will not be duplicated and never require
validation. However, when PBE is disabled, then curently creates or deletes of
instances of the OpensafImm class are accepted. That is a defect that will be fixed
separately on the supported branches. This ticket is an enhancement and is fixed
only on default(4.5).