IMM: Allow class schema change to add attribute default [#895]
This enhancement removes the restriction of not allowing a class schema change
to add a default to an attribute definition. (Note: Removing a default is still
not allowed as it is inherrently not backwards compatible.)

Only new instances of the class will get the default assigned when no value is
provided for the attribute at object-create time.

Old instances with a null-value for the attribute will still have a null-value
for that attribute even after the schema change. A general rule for schema
changes is that the immsv never alters the value of existing attributes as
part of a schema change. Doing so would require the OI to be notified, which is
not realistic since the shcema change is not performed as a CCB.

This feature is only functional after an upgrade to 4.5 has completed.