osaf: port daemonization and nid to systemd style daemon [#455] This is a first cut at getting OpenSAF to work with systemd.  This patch still leaves the opensafd SysV init.d script, but creates a new opensafd.service file which calls it.  No attempt has been made to convert the individual daemons to systemd, except for plmcd, which has been converted to a full-blown systemd daemon. built and run this for Fedora 19, OpenSUSE 12.2, and OpenSUSE 13.1. Jun 2 16:55:57 linux-po6q osafrded[16114]: Started Jun 2 16:55:57 linux-po6q osafrded[16114]: Creation of real-time thread 'OSAF_TMR' FAILED - 'Operation not permitted' Jun 2 16:55:57 linux-po6q osafrded[16114]: ER ncs_core_agents_startup FAILED Jun 2 16:55:57 linux-po6q opensafd[16086]: ER Failed Jun 2 16:55:57 linux-po6q opensafd[16086]: ER Going for recovery Jun 2 16:55:57 linux-po6q opensafd[16086]: ER Trying To RESPAWN /usr/lib64/opensaf/clc-cli/osaf-rded attempt #1 Jun 2 16:55:57 linux-po6q opensafd[16086]: ER Sending SIGKILL to RDE, pid=16106 Jun 2 16:55:57 linux-po6q osafrded[16114]: Exiting... Jun 2 16:56:12 linux-po6q osafrded[16133]: Started
Linux distros that use a version of systemd < 205 (OpenSUSE 12.1, OpenSUSE 12.2,
Fedora <= 19), fail to start opensafd because, by default, these versions of
systemd do not assign any RT time budgets to the cpu cgroups for services.

The solution is in two parts:  (1) Create an opensafd.service file for use by
systemd which has cgroup info in it.  This cgroup info is not put into the
service file if the systemd version is >= 205.  (2) Make plmcd a full
systemd-enabled daemon.

Testing Commands:
-----------------
systemctl start opensafd.service on a distro that uses systemd.
systemctl start plmcd.service on a distro that uses systemd.