ckptnd: increase performance when creating and writing large numbers ofsections [#770] When creating and writing large numbers of sections (~40k) with the ACTIVE_REPLICA bit set, on the node with the replicas (standby in N+1 config) the CPU is pegged and ckpt API functions often return with SA_AIS_ERR_TIMEOUT, including SectionCreate, ActiveReplicaSet and CheckpointClose.
The section id database is implemented as a linked list. Each write to a
section must traverse the list in order to find the section. With 1000's of
sections this takes a looong time. Also, the SectionCreate message is not
asynchronous when ACTIVE_REPLICA is specified.

Solution is in 2 parts: (1) make the section id database a C++ STL map for
fast access. (2) SectionCreate message should be asynchronous when
ACTIVE_REPLICA is specified.