amfnd : mark SU UNINSTANTIATED if all comps are UNINSTANTIATED during compfailover [#359] If a SU has one component and it moves to UNINSTANTIATED state then AMF does not generate presense state change notification for UNINSTANTIATED state of SU.
During component failover, a component moves to TERMINATING state as it is
cleaned up by AMF. This leads to TERMINATING state of SU. After successful
termination of the component it is marked as UNINSTANTIATED but SU remains
in TERMINATING state. Repair of SU leads to re-instantiation of component.
Thus component transtions from UNINSTANTIATED->INSTANTIATED-> INSTANTIATED
state. But SU is transitioning from TERMINATING to INSTANTIATED state
directly. Since presence state of SU represents combined presence states of
components, SU should also be marked as UNINSTANTIATED when a single
component is marked as UNINSTANTIATED. Same behaviour is true if all the
component are UNINSTANTIATED due to component failover.

Patch marks presence state of SU UNINSTANTIATED if all of its components are
UNINSTANTIATED during component failover.