IMM: PBE exits on errors where sqlite-handle has been closed [#996]
Errors returned from calls in the library:

   osaf/libs/common/immsv/immpbe_dump.cc

where the sqlite handle has been closed, are caught by the PBE and result
in the PBE exiting. This is a correction to the fix done for ticket #869.
That ticket altered these library routines not to exit on error, to allow
the users to clean up temporary sqlite files before process exit.

The same library functions may also fail (and close the sqlite handle)
during normal PBE processing (not during initial generation of the file).
This ticket fixes so that the PBE exits, instead of continuing to execute
with a closed sqlite handle.