IMM: Reduce log severity database size messages [#866]
The IMMND logs database size at certain threshold levels. The log severity
for these mesages was set to WArning as early as 5000 objects. The warning
severity should only be used when events occur that are unexpected or
potentially dangerous.

This changeset fixes so that:

     For every even 1000 creates a trace message is generated.
     For every even 10000 creates an INfo syslog message is generated.
     For every even 100000 creates a NOtice syslog message is generated.

In addition to this, if the number of objects exceeds 350000, then:

    For every even 10000 creates a WArning syslog message is generated.
    For every even 100000 creates a WArning syslog message is generated.

Thus warning messages are only printed when database size exceeds the
recommended maximum size.