#1009 IMM: Add checks of all reference attributes when zeroing longDnsAllowed
When the attribute 'longDnsAllowed' is changed form non-zero to zero,
this implies that the imm is transitoning from allowing long DNs to
not allowing long DNs. Such a configuration change must be validated
that the imm-db currently has no long DNs. Only if such a check passes
will the CCB modify operation assigning zero to 'longDnsAllowed' be
allowed.

The validation is done locally in the modify operation. This is sufficent
because the IMM service itself is the "OI" for the object:

 opensafImm=opensafImm,safApp=safImmService

The imm service will bar any other CCB from modifying that object untill
a current ccb that is modifying it has either comitted or aborted.

Prior to this patch, the only check done was that no object *exists*
currently with a long DN.

This patch adds a check also for reference attributes. Specifically the
attributes checked are the ones that:

 a) Are defined on the type SA_IMM_ATTR_SANAMET.
    Because currently references are only recognized as references when
    they have this attribute type.

 b) Do not have the attribute flag SA_IMM_ATTR_NO_DANGLING set.
    Because if that flag is set the reference can only point to an existing
    object, thus covered by the pre-existing check.

 c) Do not have the attribute flag SA_IMM_ATTR_RDN set.
    Because the RDN attribute is simply a copy of the RDN part of
    the objects DN, thus coverd by the pre-existing check.