IMM: saImmOmAdminOwnerClear should only be allowed for root users [#1053]
The local immnd enforces access control for saImmOmAdminOwnerClear if
access-control is enabled.

In addition, a minor change of code is done for immnd_fevs_local_checks.
The fix for ticket #938, changeset 5648:bc8d57d94f9f added a parameter 'sinfo'
to immnd_fevs_local_checks. This is a pointer parameter and it turns out
that the sinfo struct is not available to all contexts where
immnd_fevs_local_checks is invoked. This caused coverity complaints and
these where justified because the code in immnd_fevs_local_checks using
sinfo did not guard for NULL. Instead of adding such code, this patch
changes the parameter to 'uid_t uid'. The uid is what is actually used
in this function. It will be set to zero for cases where the message is
generated internally by the local IMMND. This matches the semantics of
the parameter. Even if the IMMNDs are not executing as root, they need
the root priviliges enforced by the IMM access control.