amfd: respond node level admin op at sufailover during node unlock [#1044]
When node unlock is performed on Node and it leads
to sufailover, AMFD is not responding a pending admin op.

During node unlock operation, AMFD remembers no. of fresh assignments
it has created by maintaining a counter. Whenever it gets a success response
for an assignment from AMFND it decrements the counter by 1. Thus when AMFD
gets response for all the assignments this counter will become 0 and AMFD will
respond to IMM for UNLOCK operation. In the reported issue fault during node
unlock leads to su-failover. While processing su-failover event AMFD
decrements the counter only once irrespective of the number of fresh
assignments it has created. Since counter never becomes 0, AMFD does not
respond for the admin operation.

This patch ensures if sufailover gets escalated during node unlock,
AMF will respond to IMM for the completion of admin operation.