base: Store string pointer in extended SaNameT at an aligned address [#1028]
The string pointer stored inside the SaNameT structure for long DNs (longer than
255 bytes) was not detected by Valgrind, since it was not stored on an aligned
address. This had the effect that allocated memory may have been reported as
"definitely lost" rather than "still reachable" by the Valgrind memory leak
detection.

The fix is to store the pointer at an aligned address. It will only work when
the SaNameT structure itself is stored at an aligned address (which is not
guaranteed since it contains nothing larger than an 16-bit integer), but it
should work in many cases.