plm: don't reset self during plms_ee_instantiate [#1031] Active controller resets during OpenSAF startup when modeled as an HE with dependent HEs in it.
If the active controller blade (modeled as an HE) has HE dependencies (some
other FRUs on the board), this blade/HE will only be brought in-service when
the other HEs on which it depends are brought in-service. At this point,
plms_ee_instantiate() is called and the controller ends up resetting itself.
plms_ent_enable() is called in plms_move_dep_ent_to_insvc(). This function,
plms_ent_enable(), is not called if there are no dependencies, and hardware
is not inserted. If the hardware is already present, and there are no
dependencies, plms_ent_enable() is not called, which means
plms_ee_instantiate() is not called. If the active controller blade/HE does
not depend on any other HEs, plms_ee_instantiate() is never called, and so
the blade can never reset itself.

The solution is to get the entity path from HPI, on which we are currently
running, and compare it to the HE which is the parent of the EE being reset.
If they are the same, then don't reset.

Testing Commands:

-----------------

0) you must use HPI version B.03.01 or greater to test this

1) create an HE that models the blade on which the active controller will run

2) add some HE dependencies using SaPlmDependency so that this blade depends on

   other HEs on it.

3) create an EE that models the OS for this blade

4) start opensafd and plmcd



Testing, Expected Results:

--------------------------

1) plms_ee_instantiate is called, and blade is not reset