amfd : update saAmfSINumCurr<Actv/Stdby>Assignments in SI during nodefailover [#759]
saAmfSINumCurrActiveAssignments and saAmfSINumCurrStandbyAssignment are the
two attributes of SI which together decides saAmfSIAssignmentState of SI.
When these two runtime attributes exactly matches their configuration based
counterparts, saAmfSIPrefActiveAssignments and saAmfSINumCurrStandbyAssignments,
a SI is said to be in FULLY_ASSIGNED state. Mismatch of runtime attributes with
the configuered ones leads to PARTIALLY assigned state of SI. Also if run time
attributes are 0 then saAmfSIAssignmentState will be SA_AMF_ASSIGNMENT_UNASSIGNED.

In the present problem AMF increments saAmfSINumCurrActiveAssignments when
it sends modify active assignment to a SU for an SI. This active assignment
failed with nodefailover escalation. AMF removes the SUSI for the sent assignment
without compensating for the increased count. This mismatch with the configured
values shows SI in PARTIALLY_ASSIGNED state after node joins the cluster.

This patch ensures that saAmfSINumCurrActiveAssignments or
saAmfSINumCurrStandbyAssignment are set properly in SIs during
nodefailover.