IMM: Failure to send completed to PBE defaulted to ccb-recovery [#1127]
The fix is mainly in immnd_evt_proc_ccb_apply, but also some cleanup in
immnd_evt_proc_ccb_compl_rsp where part of the same problem was addressed
by the earlier ticket: #1096.

The general case is that the ccb has just gone critical in ImmModel
(at all IMMNDs) and one IMMND is just about to send the completed
callback to the PBE. The completed callback can be sent to PBE either
in 'proc_ccb_apply' when there are no regular OIs involved with the ccb;
or in 'proc_ccb_compl_rsp' if there are regular OIs and the reply from
all of them has been received.

Solution:

The impossible case of the PBE-OI being known to ImmModel yet the
client_node not existing is handled by osafassert.

The rare but possible case of the client_node existing but being stale
is handled by skipping over the send-attempt (which would fail) and
letting ccb-recovery sort things out.

The rare but possible case of the client_node being ok but the send
over MDS resulting in an error return from MDS is also handled by letting
ccb-recovery sort it out. Note that the handling an error code from MDS
for this case as the send having been processed is safer, since the message
could possibly have reached the PBE despite that MDS reported error.

The client node stale case could in principle have been handled by aborting
the CCB, but the ImmModel has entered the critical state for this ccb at
all nodes and so such a solution would require a new message type:
revert-critical-and-abort-ccb being broadcast. Just broadcasting the
currenttly supported ccb-abort would not work because ccbs in critical
will (correctly) discard such a request.