base: remove setgid/setuid calls in execute_timed [#1138]
OpenSAF as non root cannot be started in a linux container. SMFD cannot write
its PID file.

When opensaf is running as non root in a container, the setuid/setgid feature
in ncs_os_process_execute_timed() causes problem. The new user/group is picked
up from the installed binaries. If they happen to not be the same as the
configured ones for opensaf, AMF started programs cannot write PID files.

This feature is removed also reducing the amount of system calls between fork
and exec which has been a problem.