immnd: allow service object modifications for same group [#1136]
SMF running as non root cannot change saImmRepositoryInit which is a normal
upgrade use case. Same for changing longDnsAllowed but which is one shot
change.

Reason is IMM now only allows root (uid=0) to modify service objects.

This is changed so that members of the same group as immnd, for example smfd
is allowed to change the mentioned attributes. Change of access control related
attributes is still restricted to root.