amfd: add applier support for state dump [1153]
In this patch is also added support to specify the filename to dump to
as an optional parameter to immadm. With a shared filesystem the state of
both amfds can be dump from the same node to the same filesystem.
This can be used to check for differences in state between amfds.