IMM (2PBE): Slave PBE regenerates file when restarted soon after PRT error [#1057]
If 2PBE-slave restarts in close proximity to PRT problems, then it should
regenerate the pbe-file. This is not mainly to correct for any dropped
PRT updates. That is supposed to be covered by immModel_pbeIsInSync().

It is rather to compensate for the risk that PRT problems (non OK
reply on a PRT op) are a symptom of bigger problems, that could also
compromize regular CCB consistency.