smf: Fix memory corruption and resource leak when CLM returns error [#1166]
SMFD segfaults with the following entry in syslog:

Oct 9 13:48:43 SC-1 local0.err osafsmfd[452]: ER saClmClusterNodeGet failed, rc=SA_AIS_ERR_NOT_EXIST (12)

When SMF recives a node up event from MDS, it asks CLM to provide information
about the new node. In the case CLM returns an error code (cannot find the
node), then there are a number of problems happening in the error handling code
in SMF:

First of all, free(smfnd) is called on a pointer that can either have been
returned from get_smfnd(), or from calloc(). We should only call free() when the
pointer was returned by calloc() - calling it when it was returned by
get_smfnd() will free memory that is stored in a linked list, causing memory
corruption!

Secondly, the output from saClmClusterNodeGet() is stored in &smfnd->clmInfo. In
the case where saClmClusterNodeGet() returns an error, it may be that
&smfnd->clmInfo is clobbered. It would be safer to use a temporary variable on
the stack as output parameter to saClmClusterNodeGet(), and then copy it into
the data structure only in case saClmClusterNodeGet() returns OK. A second, and
even more important reason for using a temporary variable is that the list is
protected by a mutex, but the mutex is not locked at the time of the call to
saClmClusterNodeGet().

Thirdly, we return from this function without calling saClmFinalize(), causing a
leak of the CLM handle.