IMM (2PBE): Fix incorrect transaction abort in slave PBE [#1057]
The main problem observed in #1057 is that the PBE slave gets repeatedly restarted.
The initial and main cause of the restarts is a dynamic deadlock in the slave PBE between
its main thread and the runtime thread.

<background start>
A bit of background is that the primary PBE executes all CCB and PRT jobs in the main
thread. For each CCB/PRT-job it first asks the slave PBE if it is ready to prepare the
sqlite transaction for the next identified ccb, i.e. if it has completed the commit of
the sqlite transaction for any prior ccb/prt-op AND if it has received all needed data
via its main thread for this next identified ccb. PRT-ops are identified via the "weak"
ccb-ids in the high order range.
If the slave replies ok, then it is not only rady, it has already started the prepare.

The primary PBE then prepares and commits the sqlite transaction for the CCB/PRT and
replies with OK back to imm-ram. The CCB/PRT is commited in imm-ram on reception of ok
from primary PBE (for PRTO-creates and PRTA-updates imm-ram needs ack from both primary
and slave).

Finally, for CCBs and PRTO-deletes, the slave-PBE gets the completed and apply callbacks
as applier via its main thread. PRTO-deletes are handles as full blown CCBs (but with
"weak" ccb-ids) because PRTO-deletes may be cascading, involving many deletes and are
already handled that way in 1PBE and by the primary.

PRTO-creates and PRTA-updates are handled in a slightly different way. Both slave PBE
and primary PE (and 1PBE) get only the operation-callback (create or modify). No
completed/apply callback is generated for PRTO-create or PRTA-update.
In 2-PBE the primary here also asks the slave if it can prepare and commit the PRT-op.
If the slave can prepare (if it has completed the previous job and if it has received
the matching create/modify callback via the main thread), then the slave replies ok
does the sqlite prepare AND commit. If the primary receives the ok from the slave,
it does the prepare and commit to sqlite.
Imm-ram will only commit the a PRTO-create/PRTA-update if it gets ok reply from *both*
primary and slave. Primary will only commit if it receives ok on prepare from slave
and slave will only prepare and commit if it receives the prepare request from primary.
<background end>

The dynamic deadlock of the slave occurs as follows: A regular CCB is commited, the slave
replies ok to primary, does the sqlite prepare (in the runtime thread) but not the commit.
It then goes into general poll normally to get the completed callback via main thread.
Before the completed callback arrives, the slave instead receives the PRT-op over the
main thread, the completed callback for the regular ccb is behind this PRT-op in the message
queue for the main thread at the PBE slave.

The primary-pbe has received ok from slave on the regular ccb. The primaryprepares and
commits it to sqlite, then goes back to poll and receives the PRT-op. The primary asks
the slave if it can prepare the prt-op. The Slave receives the prepare request on its realtime
thread, discovers that it is still waiting for the previos regular ccb to commit. Here we
have the dynamic deadlock.

The primary times out requestiong prepare after 3 seconds and replies with failure on the
PRT-op to imm-ram. Imm-ram aborts the PRT-op.

The slave times out after 5 seconds (in the main thread inside the prto-create/modify callback)
waiting for the prepare to be confirmed, but which is blocked by the pending commit for the
prior regular ccb, which in tern is in the in-queue for the main-thread.
The slave timeout should have unwound cleanly if there was not any bug in the slave PBE.
But here the slave PBE has the main bug. On timeout it not only discards the prt-op,
it also aborts the >>current open sqlite transaction<<.
The current open sqlite transaction is not the prt-op transaction it is the regular
ccb-transaction that is prepared and waiting on that completed callback.

When the slave pbe returns with failure) from the prt-op callback and goes into poll,
it receives the completed callback for the regular ccb on the main thread.
The protocol inconsistency is detected because it gets completed for a CCB that is gone
(aborted). This causes the slave to exit.

An additonal problem here is that the restarting slave PBE should regenerate the sqlite
file (since it never commited the regular ccb). But the slave PBE re-attaches without
regenerating the file. So besides the main bug, there is a secondary bug in the
mechanism that determines iof the slae needs to regenerate its file at restart.

The main fix done here is to avoid aborting the wrong transaction when the prt-op
fails in the slave. This was a pure bug in the PBE slave logic.

A secondary fix is to ensure that any slave PBE restart after failure to commit a
prto-op will result in regeneration of the file at the slave. This is a fix in IMMND.
Done in the next patch.

Finally some cleanup is also done, mainly in immnd_fevs_local_checks and for the
PBE to return ERR_NO_RESOURCES rather than ERR_BAD_OPERATION because the latter
is displayed as "Validation error" by the tools.