IMM 2PBE: Abort-callback before completed-callback aborts sqlite-trans [#1063]
The problem was the case where a regular ccb had entered prepare,
i.e. primary PBE has asked slave PBE to prepare the CCB.
The slave manages to prepare, including begining the sqlite transaction,
but by the time it replies, the primary PBE has timed out and replied
with failure to imm-ram. The CCB is aborted in imm-ram and abort callbacks
sent to both primary and slave PBE.

The problem is for the case when the slave PBE receives abort callback before it
has received completed, then it resets the PBE slave state variables but omitted
to abort the sqlite transaction. This resulted in the next CCB getting blocked on
the open sqlite transaction still being active:

Sep 11 12:05:30 SLES-64BIT-SLOT2 osafimmpbed: WA Sqlite db locked by other thread.
Sep 11 12:05:30 SLES-64BIT-SLOT2 osafimmpbed: WA Sqlite db locked by other thread.
Sep 11 12:05:30 SLES-64BIT-SLOT2 osafimmpbed: ER Sqlite db appears blocked on other transaction
Sep 11 12:05:30 SLES-64BIT-SLOT2 osafimmpbed: WA PBE failed to start sqlite transaction for CCB 3/3
Sep 11 12:05:30 SLES-64BIT-SLOT2 osafimmpbed: ER Exiting rc=21 (line:1140)

The slave PBE restarted and things did get resolved, but via the unnecessary escalation to
restart the salve PBE.

The fix is to add code in the abort-callback, such that the slave PBE not only resets the
state variables, but also aborts the sqlite transaction if it is active.

In addition an adjustment is made so that the PBE does not linger in a local wait loop
for the case where the problem is a bussy sqlite transaction at the slave. Instead a
TRY_AGAIN reply is generated immediately towards the primary PBE. The only point of this
second change is to speed up the resolution of such a blockage, in case it should happen.
But it should really never happen. If the state variables are zeroed there should be
no active sqlite transaction in the slave.