IMM: Sync-retry reverted back to milliseconds with increase on retries [#1188]
The sync retry time was changed from 150msec to the huge value of 2 seconds
in ticket #851 (OpenSAF 4.5). This large value may have been apropriate
for optimizing some variant of sync (?) but it is not optimal in general.

The retry time is reverted back to milliseconds level, but with the value
increased with each retry up to a max retry time of 0.5 seconds.