IMM 2PBE: Slave main thread avoids aborting ccb during prepare [#1080]
The incident described by this ticket has some similarities to #1063 in that
it is caused by an early abort of an on-going ccb. The problem in #1063 was
caused by the omission to abort the sqlite transaction in the slave PBE.

In this case (#1080) we see a segv crash of the slave PBE. The crash is caused
by the abort-callback arriving in the main thread of the slave PBE concurrently
with the runtime thread in the slave PBE still being bussy performing the
prepare of that CCB. The prepare in the runtime thread is iterating over the
immutils collection for the CCB at the same time as the main-thread deletes
that collection. This causes the segv in the runtime thread and the slave PBE
restarts.

The fix done here in the abort-callback of the main-thread, is to detect
the case of this being the slave PBE, where prepare has started but not yet
completed for this same CCB (by the runtime thread). If this is the case,
then the main thread (abort callback) of the slave will wait for up to 3
seconds for the prepare to be completed by the runtime thread. If total wait
expires with prepare still on-going, then the slave PBE main thread will
trigger exit for the slave PBE process. The slave PBE will then restart.

Note that this whole case is an error (to be fixed) in an error case (abort
of a ccb during prepare). Normally the prepare should complete in under 3
seconds and the abort callback will remove the prepared CCB safely.

For normal CCB flow, where the primary PBE has received ACK from slave that
it has started prepare, there is no upper limit on the time a prepare can
take. The early abort callback that interferes with the CCB could only be
caused by the primary PBE not receiving reply from slave PBE in time on the
prepare request (from primary to slave). When the slave PBE finally starts
executing the prepare, the abort of the CCB has already been decided by the
primary and imm-ram.