smf: handle the suspended single step cluster reboot campaigns to continue after reboot [#1118]
This special case happens with single step cluster reboot campaigns, only when the campaign is suspended at the time when the cluster reboot step is executing.
The suspending event remains pending and only processed after the current procedure returned.
This never happens with a cluster reboot step, since the cluster is rebooted and the event is not processed. The campaign state remains suspending and the procedure state remains executing.
This fix handles this special case after the reboot.
2 possible outcomes:
   -campaign finishes, since no more procedure left to execute.
   -remain in suspend state, since procedures left to execute.