2PBE: PBE slave restart during class-create must regenerate imm.db [#1092]
The imm.db file is regenerated by the slave if cb->mPbeVeteranB is false.
The logic for setting mPbeVeteranB to false has been corrected in a few
places to be more eager. In particular inside immnd_forkPbe If PBE in
general (either primary or slave or both) is not in sync just before a new
PBE primary or slave is forked, then they should always regenerate the imm.db
file. This workeed correctly for the primary PBE but not for the slave PBE,
before this patch.