IMM: saImmOmCcbValidate must cancel ccb-timer before returning [#1201]
saImmOmCcbValidate retuurned without canceling the ccb-oi-timer resulting
in the potential abort of the ccb due to timeout when the om-user is
idling, between saImmOmCcbValidate and the termination of the ccb.
The ccb can be terminated by either saImmOmCcbApply or saimmOmCcbAbort.
In the timeout case, the saImmOmCcbApply will return ERR_FAILED_OPERATION
signidying the ccb was aborted, even if validation succeeded.

This patch fixes so that the timer is canceled before returning to the
om-user from saImmOmCcbValidate.