IMM: Correction to idempotent ObjectImplementerSet during sync [#1202]
The case of an object-applier (note *not* class-applier) re-attaching
idempotently (it already has existed as an object applier for that
object on the local IMMND) was not handled correctly. This caused
the function to return OK when it should have returned the special
return code SA_AIS_ERR_NO_BINDINGS which here has the internal
meaning of idempotent OK.

In addition the name of the function 'ImmModel::setImplementer()' is
changed to 'ImmModel::setOneObjectImplementer()' to make its role clearer
and avoiding confusion with the function 'ImmModel::implementerSet()'.