smfd: no unneccessary delay at node reboot [#1203]
After SMF have rebooted one or several nodes, it polling for "node UP" events.
A new poll was made every 5 sec. A fault make it wait for 5 sec even if all nodes was reported UP.
Same fault when trying command "true" in the nodes reported to be UP.
This patch removes the unneccessary delays and decrease poll time to 1 sec for "node UP".