IMM: Detect long DN/parentDN, long RDN, long SaNameT attr for create callback [#1204]
The IMMND server will screen the contents of every ccbObjectCreate for
long DNs or long RDN. If any long DN or RDN is detected, then the message
sent to the client (for the callback) will be a new (node local) message type.
The new message type conveys the same semantics as the old create-callback
message type but also flags that the callback has values that are not
compatible with any OI/applier that is not long DN capable.

The imma library uses this info to check if the client is long DN capable.
If the OI/applier is not capable. the lib does not generate the create
callback. Instead the library bounces internally with error directly to the
server. The CCB will fail (get aborted). For appliers, the dispatch loop
is broken and the handle is invalidated. This is necessary because the applier
can not reply to the server. IF the main OI is long DN capable, or there is
no main OI, then the CCB can succeed in commit, yet the create callback never
reaches the applier. To avoid a silent breach of contract with such an applier,
the handle is instead stale-marked and BAD_HANDLE returned to the applier.

This patch also adds a check that prevents long DNs from being created
before upgrade to 4.5 is complete (protocoll45Allowed()). The risk is small
but in theory some application could try to generate a long DN object
while the upgrade to 4.5 is in progress. Without this added check, the
create could succeed on upgraded nodes and fail on not yet upgraded nodes,
casuing the imm database to be inter-node inconsistent.