IMM: Protect non long DN capable clients in modify callback [#1204]
Same fix as for the create callback but here for the modify callback.
The RDN issue is not relevant for the modify callback because the RDN
attribute will never be part of such a callback, it can not be modified.

The IMMND server will screen the contents of every ccbObjectModify for
long DNs. If any long DN is detected, then the message sent to the client
(for the callback) will be a new node local message type.
Two cases need to be caught.
a) A modify of an SaNameT attribute trying to set a long DN.
b) The object itself may have been created with a long DN whith no
OI attached. Thus we have to detect the long DN case also. This is also
relevant to protect incapable appliers even if the OI is capable.

The new message type conveys the same semantics as the old modify-callback
message type but adds the information that the callback has values that
are not compatible with an OI that is not long DN capable.

The imma library uses this fact to check if the client is long DN capable.
If not it does not generate the callback. Instead the library bounces
internally with error directly to the server. The CCB will fail (get aborted).
For appliers, the dispatch loop is broken and the handle is invalidated.
This is necessary because the applier can not reply to the server. IF the
main OI is long DN capable, or there is no main OI, then the CCB can succeed
in commit, yet the create callback never reaches the applier. To avoid a
silent breach of contract with such an applier, the handle is stale-marked
and BAD_HANDLE returned to the applier.