IMM: Protect non long DN capable clients in delete callback [#1204]
Same fix as for the create and modify callbacks but here for delete-cb.
The only thing relevant for the delete callback is the object DN.

The IMMND server will screen the DN of every ccbObjectDelete for having a
long DN. If this is detected, then the message sent to the client (for the
callback) will be a new node local message type. This is also relevant to
protect incapable appliers even if the OI is capable.

The new message type conveys the same semantics as the old delete-callback
message type but adds the information that the callback has a DN that
is not compatible with an OI that is not long DN capable.

The imma library uses this fact to check if the client is long DN capable.
If not it does not generate the callback. Instead the library bounces
internally with error directly to the server. The CCB will fail (get aborted).
For appliers, the dispatch loop is broken and the handle is invalidated.
This is necessary because the applier can not reply to the server. If the
main OI is long DN capable, or there is no main OI, then the CCB can succeed
in commit, yet the delete callback never reaches the applier. To avoid a
silent breach of contract with such an applier, the handle is stale-marked
and BAD_HANDLE returned to the applier.