amfd: adjust rank of all csis in SI if csidep is deleted [#1240]
Before unlock and unlock-in operation, user deletes CSI dependency via CCB
modification. After this when unlock-in and unlock operations are performed,
AMF did not give callback to components for some CSIs.

While deleting CSI dependency, AMF assigns rank 1 to the deleted CSI in SI
if this CSI has no further sponsor. But AMF is not cheking if this CSI
is sponsor to some other CSI recursively. Due to this there was no CSI with
rank 2 in this SI when this CSI of rank 2 was deleted. Now AMFND assigns CSI
of rank r+1 when csi of rank r is assigned. Since there was no CSI with rank 2,
AMFND skipped assignments when CSI of rank 1 was assigned assuming there are
no further CSIs to be assigned. Thus CSI of rank 3 remained unassigned. So deletion
of CSI dependency requires recalculation of ranks of all the CSIs in SI.

Patch fixed the problem by performing recalculation of ranks of all CSIs in a
SI when CSI dependency is deleted through CCB modification.