test-mds: tune testsuite 6 of 6 message size below the default heartbeat time ofavsv [#1005] test-mds: tune testsuite 6 of 6 message size below the default heartbeat time ofavsv [#1005] V2 Bug : mdstest sometimes fails when running tcp Analysis : Test case tet_broadcast_to_svc_tp_6() of mdstest suite 6 has test case of sending Broadcasting above fragment size (>MDS_DIRECT_BUF_MAXSIZE - 65479 ) as a VERY HIGH Priority message.
While sending this message, the Opnesaf heartbeat message which are being
sent in periodic time , which are configured with default time of avsv is
getting blocked for a while , that is why we are seeing the problem.

solution : we have tow option to provide solution

 1) While running the mdstest tune the avsv heartbeat timeout values to
higher. 2) Or tune the mdstest Broadcasting VERY HIGH Priority message Size
to lesser to match the default heartbeat timeout values of avsv.

I fixed the issue with second option.

Also fixed some wrong parameter passing mds_sendrsp_getack() in test code.
Done some memsets()