amfd: mark unlocked-in amf node disabled upon clm exit [#496].
An unlocked-in AMF Node was not disabled, when CLM lock was done on the mapped
node.

AMF does not conform to AMF B.04.01 spec: Section 3.2.6.2:
If a node is enabled and not in the locked-instantiation administrative state
when it leaves the cluster membership, the node becomes disabled while it is
out of the cluster and becomes enabled again when it rejoins the cluster.

Patch marks AMF node disabled if it is enabled and in unlock-in state
at the time of CLM exit.