amfnd: issue remove cbk for assigning csi after successful assignment V2[#1046]
This is version-2 after removing check on SI for REMOVED state.
This check is not needed as a SI moves to REMOVED state only
after all of its CSIs are removed.

A user locks a SI and a component faults with comp_restart recovery during
remove callback on standby SU. After this when user tries to unlock SI,
unlock operation gets timed out.

When AMFND gets removal of assignment on standby SU (having three comps),
because of lock operation on SI, it issues removal callback to comp1. While comp1
is handling removal callback another healthy component, which is still assigned
for a CSI of this SI, faults with comp_restart recovery. When comp2 gets repaired,
AMF reassigns CSIs to it. In the meantime, when comp1 responds for the removal callback
AMFND does not issue removal callback to the CSI of comp2 as it is in assigning state.
AMFND goes on and finishes removal in comp3. After successful reassignments in comp2,
AMFND does not issue removal callback to comp2. Due to this, AMFD
did not get response from AMFND for successful removal of SI.

Patch fixes the problem by issuing removal callback for a assigning CSI
after successful assignment.