ntfa: fix mutiple mutex lock/unlock on ntfa_cb in saNtfNotificationSend [#1263]
In a multithreaded application,saNtfFinalize() may hang forever
leading to deadlock, if one thread tries to call finalize the handle,
when other thread is trying to send a notification.

The reason is saNtfNotificationSend() API takes mutex_lock() on ntfa_cb three
times:
-In the first lock/unlock,it retrieves client record.
-In the second lock/unlock, it verifies notification parameters.
-In the third lock/unlock, it sets notificationId that it has got
  from ntfs.
In saNtfFinalize() API also after receving response from the NTFS, it
tries to free all records for this client and also it releases allocated
memory for the notifications.
If saNtfFinalize() is called for the same client from one thread for which Send() is in
progress in another thread, saNtfFinalize() may get lock on ntfa_cb and it will try to
destroy all the records for this client. But here it will not be able to destroy
the record as Send() API is still using records in the lock free region.
Thus after  taking a lock on ntfa_cb, saNtfFinalize() blocks itself for the client
record to be released by Send(). Since Send() is in progress, in between it will also
try to again take lock on ntfa_cb, but it will not get as it already owned by saNtfFinalize()
which is already blocked because of client resources owned by Send().

As a solution, multiple locks should be avoided in the Send() API. Only while retrieving
client record and notificationHanndle record it should use lock on ntfa_cb.
The reason is saNtfFinalize() will anyway wait for the resoureces taken by Send() to
be released and thus it will not free the memory. In the worst case, saNtfFinalize()
will complete in NTFS_WAIT_TIME, but there will not be any deadlock situation.