IMM: Larger timeout on cleanup of PRT/class-change request continuations [#1266]
The IMMND server maintains request continuations containing the destination
for replying on non CCB requests such as to update PRT data or class-create,
schema-change and class-delete.

Some of these requests can take more time to process than is normal for an
IMM request. A class-change can result in instance migration and there may
be many instances. While there should only be one or two sqlite transactions
for the entire procedure, the number of sqlite commands can be large.
This may take time to process (up to a minute or more has been observed) and
the result is then that the user gets an ERR_TIMEOUT on the request.
If the "user" is a shell script or program, then the ERR_TIMEOUT is typically
treated as a failure to perform the operation, resulting in a failure of
an upgrade campaign. ERR_TIMEOUT can actually not be interpreted as failure
to execute a request. The true meaning is that it is unknown if the request
has failed or succeeded. Typically it is still being processed when the client
gets timeout.

The client can increase the API level syncronous timeout by setting the
environment variable IMMA_SYNCR_TIMEOUT. But the problem here is that the
IMMND server has an internal timeout of when to garbage collect the
continuation records it keeps for pending replies. THe IMMA_SYNCR_TIMEOUT
does not control the server internal timeout.

This patch increases the hardwired server side timeout for cleanup of
PRT equests, from 6 seconds to 120 seconds.