IMM: Halt sync with SIGKILL insted of SIGTERM and replace assert with error [#1295]
The problem was triggerde by the handling of the unusual case of a PRTA
update failing during on going sync due to PBE being configured but
unavailable at that time. Such a failed PRTA update has to abort the sync
becasue the PRTA update may succeed at sync clients yet fail at veterans.
This is due to the sync clients possibly not yet having synced the config
data specifying if PBE is available or not.

The handling of this case by the IMMND coord was to send a SIGTERM to the
sync-process and to broacast an abort of the cluster wide sync job over fevs.
A SIGERM may however be slow to get processed by the sync process. If the
abort of the sync job gets handled before the sync process has terminated;
and the sync process sends additional sync messages arriving at a processor
after the sync has been aborted, then the assert will get triggered.

The solution is to replace the SIGTERM with SIGKILL. THat should drastically
reduce the probability of the sync process lingering after the sync job has
been aborted in the IMMND cluster. In addition the assert is dissarmed and
replaced with a return of error (to the sync process) for this case.