immsv: Revert imma_client_node->replyPending to unsigned char [#1341]
The member imma_client_node->replyPending was in OpenSAF4.1 and earlier
defined as 'uns8', i.e. a byte. But this member is a counter and not a boolean.
The uns8 type was often used as a proxy for a boolean type in OpenSAF 4.1 and
earlier. This was before 'bool' was available as a first class type for gcc.
At some point in OpenSAF 4.2 the bool C type became available in gcc. There
was then a general sweep across all services to change the use of uns8 to bool.
The replyPending member was swept along in this change, but code actually
still increments and decrements the value, which makes no sense for a boolean
type.

The intent of the replyPending member is to keep track of outstanding
replies on requests from imma library to immnd server using the handle.
This so that a restart of the local IMMND will result in the handle being
marked as 'exposed' and not resurrected.
An IMMND restart will mean that any unreplied requests will have lost
their replies. This violates the interface contract from the imm service
side towards the client using the handle and so the handle must not be
allowed to get resurrected. Instead the handle must be marked as exposed.
The client will then get an ERR_BAD_HANDLE from either saImmOxDispatch
(failed active resurrect) or from the next syncronous downcall made after
IMMND went down (failed reactive resurrect).

For syncronous requests, the count will only go from 0 to 1 and back to 0 on
reply. This ticket does not affect syncronous requests. For asyncronous
requests it is possible for the client to invoke more than one request,
before entering poll to receive replies. For asyncronous requests the
replyPending member must work as a counter and not a boolean. The effect
of being a boolean is that a handle may get resurrected when there is
still asyncronous requests unreplied to, i.e. the replies would get silently
lost.

This changeset restores the type of the repliesPending member to unsigned char.