smfd: get nodes for activation without warning if entity was a node [#960]
Symptom: Notice in system log saying that no nodes found for DN. This
message would show when using nodes in activation actedOn tag in a campaign.

Analysis: Just a harmless notice, nothing else seems to go wrong. Even the
multiple procedures went fine

Solution: Getting nodes for activation no longer produces warnings if the
entity specified in the activation is a node. The used function returns a empty
string if input DN is not of expected class