amfd: ignore locked-in node during lock, shutdown and unlock op on ng [#1356]
If a node is in lock-in state and lock, shutdown or unlock are initiated on any of its
nodegroup then there are two problems:
1)During lock and shutdown operation on ng, locked-in admin state of node is changed
to locked.
2)During unlock operation on ng, locked-in admin state of node is changed to unlocked.
In both the above cases locked-in state of node should remain intact.

Patch ensures that a locked-in node remains untouched during lock, shutdown and unlock
operation on any nodegroup of node.