immtools: audit no dangling references in PBE file [#19]
The patch covers only aditing no dangling references. It's required by critical ticket #1377.
The patch first check no dangling references for multi-value attributes, then it checks no dangling references for non multi value attributes.
This is done in two steps because values for multi value attributes are stored in another table.

The patch is done in libs/common/immsv, so that it can be used by IMM PBE in later implementation.
If the audit fails, one line is written in the console. Other auditing details are logged in syslog.