logsv: Improve memory handling when writing log records [#1376]
Once having request to write log record, LOGSV allocates
three separate memories and do copying log record to there.
As the result, it consumes memory and decrease performance
if log record is in big size.

With this fix, log record memory is allocated by the caller and this memory is
used or referred during log writing process including time-out/normal cases.

Note that, In most cases, the caller thread is blocked until the log handler thread
finishs the request (e.g: writing log record to file).
Therefore, when the caller thread is unblocked, it is safe to free the log record memory
as the log handler thread no longer uses it.

But in time-out case, it is unsure when the log handler thread
have done use the log record memory. So, in worse case, if the log record memory
is freed by the main thread while the log handler is using it, it will cause the problem.

To make sure there is no corruption of memory usage in case of time-out,
We leave the log record memory free to log handler thread.

It is never a good idea to allocate and free memory in different places.
But consider it as a trade-off to have a better performance of LOGsv
as time-out occurs very rarely.