build: Use the C11 standard for C and the C++11 standard for C++ [#1374]
Use the C11 standard when compiling C code, and the C++11 standard when
compiling C++ code. For GCC, this means that we also bump the recommended
minimum version to 4.8.1, which is the first version where all major features of
C++11 were fully implemented (though most C++11 features were implemented
already in earlier versions).