amfd: fix N+M identification of standby SU at controller failover [#1414] An N+M SU, with an active assignment, can also get a standby assignment at the same time. Details for reproducing this are in the ticket.
The SUSI returned from avd_sg_npm_su_othr() is not being checked to make
sure that its HA state is standby (among other things).

The SUSI returned from avd_sg_npm_su_othr() needs to check that its HA state
is standby, that its readiness state is in-service, and that its fsm state
is not unassigned. This is done in many parts of sg_npm_fsm.cc, but it is
missing in some parts.