clm: don't always checkpoint nodes in plm readiness callback completed state [#1416] At active controller switchover, the newly active sends node down events for nodes that are not down.
There is a race condition on the standby between receiving the node_down
checkpoint message, which removes the node from the node_id database; and
receiving a checkpoint from PLM readiness callback COMPLETED state. This can
cause an entry to be created on the node down list, and never removed.

Don't checkpoint nodes in PLM readiness callback, if node->nodeup == 0 and
state is COMPLETED. In this case we know a node down checkpoint message has
been sent which removes the node from the node_id db on the standby.
Checkpointing here can readd it, which we don't want to do.

Testing Commands:
-----------------
1. 3 or more nodes in the cluster, all with EEs
2. reboot one or more of the payloads, wait for them to come back up
3. perform si-swap of controller

Testing, Expected Results:
--------------------------
1. clm should not issue node down events for the nodes that were rebooted