amfd: don't crash if nodegroup is added with node in the same CCB [#1458]
Currently, if a nodegroup and a member node is created in the same CCB, and
the nodegroup is modified to include this new node, amfd will crash.

This change will reject the change instead of trying to deference a null 'ng'.
Perhaps support for this use case can be added later.