amfd: maintain runtime updates for su, comp, si and csi at standby [#1141]
Amfd was killed when lock operation on su was going on.
After successful failover, susi of locked su is still shown.

In the reported issue, lock operation was successful and active AMFD was able to
delete SUSIs and COMPCSIs in its database before got killed. AMFD was killed
when it was performing run time updates for su, comp, si and csi to imm.
SUSI and COMPCSI of locked Su is still shown because active AMFD was killed
without performing all the updates and since these updates are not maintained
at standby AMFD, even after failover imm was not updated for the same.

Patch proposes a solution, in which standby AMFD will maintain run time
updates for su, comp, si and csi. However the size of the job queue will not
execeed more that 200. Job queue will be emptied iat standby if it crosses this max size
limit. The value of 200 is taken from the reproted issue in which AMFD
missed around 150 updates to imm for comp, su, csi and csi during failover.

TODO:
This solution has following limitations:
1)Fix size(200) of job queue at standby.
2)Standby maintains updates for only su,comp, csi and si.
3)New active may still miss some updates if updates are more than 200.

Improved solution will be something like this:
1)Standby AMFD will maintain updates for all the classes and with no size
limitation.
2)Whenever active controller is finished updating to imm and there are no
more jobs to update, it will ask standby AMFD to flush its job queue.

Implementaion of this will require new messaging (AMFD version update) and
thus will be backward incompatible.