plm: fix deactivate for ATCA [#1289] When SA_PLM_ADMIN_DEACTIVATE is issued for an ATCA FRU, the HE presence state gets stuck in SA_PLM_HE_PRESENCE_DEACTIVATING, and never transitions to SA_PLM_HE_PRESENCE_INACTIVE.
The act_in_pro and deact_in_pro variables for the entity are not
cleared. So, when the HPI hotswap state event occurs which says that the
entity has been deactivated, PLM never transitions the state to INACTIVE.

act_in_pro and deact_in_pro need to be cleared before setting
SA_PLM_HE_PRESENCE_DEACTIVATING and SA_PLM_HE_PRESENCE_ACTIVATING,
respectively, so that when we receive the HPI hotswap state event, the FRU
can be deactivated or activated.

Testing Commands:
-----------------
1) Issue SA_PLM_ADMIN_DEACTIVATE to a locked Full-5 Hotswap HE entity


Testing, Expected Results:
--------------------------
1) HE entity should transition to INACTIVE, and not get stuck in DEACTIVATING