amfd: do not assign SUs on locked nodes when ng is deleted. [#1507]
A nodegroup is deleted in locked state when some of the nodes are kept in
admin unlocked state. Application remains in unassigned state.

In CCB apply for deletion of NG, AMF is going for new assignments on the
nodes of NG which are in locked state.

Patch fixes the problem by calling the assignment logic only for unlocked
nodes in CCB apply for deletion of NG.