amfd: reject si-swap if SI will remain with only active state (NWAY)[#1516].
In the given configuration, swapped SI2 was fully assigned in the beginning
but it had only one standby assignment. When SI-SWAP operation was performed
on the SI, it caused removal of quiesced state after standby took the active
role.

In N-Way model a SI will have multiple standby assignments, so even if swap leads
to removal of one standby there would be other standby SUs to take over the active role.
So, I think, a general check for checking saAmfSGMaxStandbySIsperSU count for
highest ranked SU is not needed and operation should be allowed if there are
more than two standbys SUs exist for the SI and it will lead to removal of
assignment after swap. A check can be added in a case where SI got reduced
(by configuration or due to degraded state of the system) to 2N type of redundancy
and operation will leave SI with only active state by removing the quiesced state.

Patch rejects si-swap if there is only one standby assignment for SI and
si-swap will lead to removal of quiesced assignment becuase of
saAmfSGMaxStandbySIsperSU constraint.