smfd: renumber proc action Id when merge procedures [#1600]
Whitout the patch, ff two procedures which will be merged address
the same IMM object for creation/deletion/modify in the procInit
and/or procWrapup, multiple rollback element may receive the same DN
causing the campaign to fail.

Solution: in merging method, renumber the Id of the procInit and
procWrapup actions.