ntfd: fix syslog warning when queueing notification for relogging [1628]
syslog is flooded with the following warnings
Dec 14 12:25:04 SC-1 osafntfd[9377]: WA Error when logging (9), queue for relogging

Ntfd logs notifications by calling async api of LOG service. If ntfd gets error in the
confirmation callback, it queues up notification for relogging and logs a syslog warning.
If LOG service is busy for longer time for example in failover situation, then it may give
TRY_AGAIN to ntfd. In such a case syslog warning can be avoided and can be logged as trace.


Ptach converts warning in a trace irrespective of error code.