imm: Canonicalize attributes presented by OiCcbObjectModifyCallback [#801]
If the modType is SA_IMM_ATTR_VALUES_ADD or SA_IMM_ATTR_VALUES_DELETE,
it will be converted to SA_IMM_ATTR_VALUES_REPLACE and the values of the attr-mod will be the values in the after image.

This feature is only for OIs with version A.2.17 or later so we don't modify the content of incoming IMMND_EVT_A2ND_OBJ_MODIFY messages.
New memory is allocated for canonicalized attribute-modifications instead.

PBE will also receive canonicalized attMods.

OIs and appliers which are initialized as A.2.17 or later MUST support SA_IMM_ATTR_VALUES_REPLACE in their callbacks.