amfnd: terminate npi comp in non-restartable SU during admin SU RESTART #1561]
First user performs restart admin operation on SU.
After this shutdown operation is performed which left SG in unstable state.

Configuration used in the issue configures compinstantiationlevel for a
NPI component and next instantiantionlevel belogns to a PI comp. Since a NPI
component is instantiated when it is assigned some work load.
Issue is reproducible when a configuration meets following criteria:
-Atleast one non-restartable component is present in the configuration.
-All non restartable component should be unassigned.
-NPI comp should be restartable.
-CompInstantiationlevel is configured in such a way that last
instantiationlevel belongs to a PI component which is non-restartable and
unassigned.

When SU restart operation is invoked, AMFND picks the last
component which is an unassigned non-restartable PI comp and starts terminaing
it (avnd_su_pres_inst_surestart_hdler). After successful termination, next
component is a NPI componenti.AMFND skips it and starts terminating next PI comp
(avnd_su_pres_terming_compuninst_hdler). When SU moves to Instantiated state and
AMF tries to reassgin SIs, SI assigned to NPI comp remains in Assigning state as
AMFND could not instantiate the component (it was not terminated).
After this when user tries shutdown operation on SU and AMFND sends quiescing
state to AMFND, this assignment gets postponed as AMFND waits infinitely for
previous assignment because of surestart is still pending.

Patch fixes the problem by terminating the NPI comp in a non-restartable SU
during RESTART admin op on SU.