clm: fix node_down handling at standby when preceded by admin op [#1643] Node downs are checkpointing to the standbys. Standby CLMD is also a subscriber of MDS NODE_DOWN events. The Standby CLMD should ensure integrity when processing duplicate or out-of-order arrival of CHECKPOINT message for NODE_DOWN and MDS NODE_DOWN events. This ticket fixes a scenario where a node-down is not processed because the node is already marked as a non-member via an admin op that preceded the node's restart.
clm: fix node_down handling at standby when preceded by admin op [#1643]
Node downs are checkpointing to the standbys. Standby CLMD is also a
subscriber of MDS NODE_DOWN events. The Standby CLMD should ensure integrity
when processing duplicate or out-of-order arrival of CHECKPOINT message for
NODE_DOWN and MDS NODE_DOWN events. This ticket fixes a scenario where a
node-down is not processed because the node is already marked as a non-
member via an admin op that preceded the node's restart.

Testing Commands:
-----------------
Perform the following sequence continously

1) Lock a node
2) Restart the node
3) Switchover

Testing, Expected Results:
--------------------------
There should not be any stale notification marking a node as non-member.