clmd: Fix coredump in handling node_up message due to null ip [#1647] The coredump happens if the node_up is sent from clm node agent not attaching the ip. The error code is set to SA_AIS_ERR_NOT_EXIST, but it's incorrectly overwriten to SA_AIS_OK, therefore ip is null though being referenced to in later part of proc_node_up_msg.
clmd: Fix coredump in handling node_up message due to null ip [#1647]

The coredump happens if the node_up is sent from clm node agent not
attaching the ip.

The error code is set to SA_AIS_ERR_NOT_EXIST, but it'sincorrectly
overwriten to SA_AIS_OK, therefore ip is null though being referenced
to in later part of proc_node_up_msg.

The patch ensures the error code is not overwriten to SA_AIS_OK if it
finds error due to null ip