amfnd: Simplify if statement for qualified assignment component [#1538]
There's a compound *if* conditions to determine whether a component
can be assigned/removed for a csi. Duplication of it has been placed
in some amfnd functions. Also, it's not quite informative since the
false case is being logged in generic meanings.

The patch makes a separated function to do this valuation, add more
traces for specific cases