amfd: fix assignment of standby HA state in NWAY model[#1523]
In the application, compcstype is not configured for csi of one si(say SI6).
Unlock-in and unlock of sg leads to only active assignments for Si execpt
SI6.

In assignment function avd_sg_nway_si_assign(), AMF tries to first assign
active HA state for SIs. If atleast one active assignment is given, standby
assignments are skipped. When all the active assignments are over same function
tries to assign standby HA state. This decision is taken based on one of the flags
is_act_ass_sent. Every time above function is invoked it tries to issue active
assignments by setting flag is_act_ass_sent to true and then it creates SUSI.
But this flag is not set to false when no SUSI (assignment) is created successfully.
In the issue, all SIs execpt SI6 are assigned active HA state. In SI6, CSIs could not
be assigned because compcstype is not configured for its CSI. So the flag remains true
even though no SUSI for SI6 was created. This function is invoked two times a)when SG
is unlocked and b) when AMF gets active assignment responses for other SIs.
Since SI6 is unassigned, function will always try for active assignments for SI6 and flag
is_act_ass_sent will remain true in each invocation of this function. Because of this
function will always return wihtout trying for standby assignments for other SIs.

In the same configuration if one tries to lock and unlock SIs other than SI6, then
lock operation will fail because avd_sg_nway_si_assign() will try for assignments
for both locked SI and SI6. Failure of active assignment for Si6 will change the return
status of function which, eventually, results in failure of lock operation.

Patch fixes the problem to set is_act_ass_sent true only when SUSI for atleast one SI
is created successfully. Also patch fixes the problem related to SI lock by checking
the SUSI creation for the SI.