pyosaf: Fix Ccb and ImmObject classes to handle more inputs [#1663]
Fix the Ccb class to allow entering values as atoms in modify
operations when there is only a single value affected.

Fix ImmObject to automatically prepend <rdn-attribute-name>= to the
RDN attribute value to make it adhere to the format expected by IMM,
if the user only provided the actual value.