amfnd: fix amfnd crash due to inst failure of a restarting npi comp [#1587]
amfnd crashes when comp instantiation fails as a part of comp restart recovery.
Applicable to a NPI SU only.

Crash occurs because AMFND is trying to respond to AMFD for assignment status.
If a healthy comp faults in stable state of SG with comp-restart or su-restart
recovery and it results in INST_FAILED state of SU, then there will be no pending
assignments to AMFD. AMFND must inform to AMFD only when SU enters INST_FAILED state
during handling of assignments sent by AMFD.