imm: add IMMD and RPM config changes, and new XML schema for cloud resilience feature support [#1625]
New option for enabling/disabling cloud resilience feature is added to immd.conf
With the new feature, IMMND and IMMD need to be installed on payloads. RPM spec file is updated to install IMMD and IMMND on all nodes.
New XML schema is added with support new attributes needed for cloud resilience feature