log: add support for cloud resilience feature (test part) [#1179]
The patch makes LOG service be able to handle the case that both SC nodes
are down at the same time.
When one or both nodes go up again the log service must be able to
resume its work preferably without actions by the clients.

A log client should not have to be aware of if one or both SC nodes are down.
The only thing that should happen is that a TRY AGAIN (and in some cases TIMEOUT) returned.
It is the responsibility of the client to decide how to handle this.