amfd: do not checkpoint the nway sg fsm in case of no change [#1697]
After the nway application AMF entities are created, active amfd calls avd_sg_nway_si_assign
function to assign any unassigned SI but all SUs are locked so no SI are assigned
and the FSM is still stable. In this function, it always checkpoints
the nway sg fsm with state STABLE (at the beginning of the function) to standby
even the fsm state is not changed.
When the standby amfd receives the checkpoint message, it founds that the nway sg doesn't exist
in its database and does an assert to generate the coredump. The nway sg doesn't exist in standby
is because checkpoint message arrives before the ccb_apply_ccb (which informs the nway sg creation)
from immnd.

Object's checkpoint message arrives to standby when the corresponding object doesn't exist is a rare
and race condition case, it needs a general solution.

But to solve for this particular crash, it is changed to not checkpoint the nway sg fsm state as
it's not changed (this is the correct behaviour and aligned with other sg).
Since nway sg fsm is not checkpointed because of no change, crash will not happen.