lcknd: fix lock sync bug during election of new master [#1488]
Exclusive locks which are held when the master nd reboots can never be
unlocked, and pending exclusive locks which are held when the master nd
reboots can never be given out.

When the master nd reboots and a new master nd is elected, the other nds
which have locks in use resend the lock information to the new master. But,
not all information needed by the new master is sent. "lcl_resource_id" and
"lcl_lockid" are not sent. So, when the new master sends an unlock response
to an nd which was in the middle of unlocking with the old master, the nd is
unable to find the local lock because the master didn't send all the right
data. Same when the new master send a lock response to an nd with a pending
lock.

The solution is in three parts. (1) lcl_lockid was not being put into the
GLSV_LOCK_REQ_INFO message by MDS, even though it was part of the message
structure. This is now done. (2) add lcl_resource_id to the
GLSV_EVT_GLND_RSC_INFO message so that the new master nd can properly set up
the lock queues. And (3) unlock requests sent to the master nd which has been
rebooted, need to be handled by the new master nd, and unlock responses sent