amfd: Add saAmfUnassignedAlarmStatus attribute to memorize the alarm_sent status [#1620]
If the SI Unassigned Alarm is raised before headless by locking SU for instance, then
after cluster recovers from headless and unlocking the SU, this alarm is not cleared.
As the application can reside in PL nodes and it's right to expect the previous raised
alarm should be cleared once the SI gets back assignments. The patch adds new attribute
saAmfUnassignedAlarmStatus attribute to SaAmfSI class to memorize the variable alarm_sent
for headless.