rde: Generalize the role selection algorithm to work with more than two nodes [#79]
Without changing the network protocol, generalize the role selection algorithm
used by RDE so that it works with more than two nodes. Nodes are now initially
given the QUIESCED role. When there are no ACTIVE or STANDBY nodes, the CLM node
agent will initiate an election of a new ACTIVE node. RDE on the winning node
will set its role to ACTIVE.

There is now also a new customizable script which is executed on a node before
it becomes ACTIVE. The purpose of this script is to perform necessary
preparations (if any) needed before a node can take on the ACTIVE role. For
example, shared files system may have to be mounted.