ckpt: Support AMF configurations containing more than two OpenSAF 2N SUs [#79]
Add support for configuring the system with more than two OpenSAF 2N SUs. In
particular, this means that all OpenSAF directors must support starting up and
running without (initially) getting any assignment from AMF. The directors on a
spare system controller (i.e. a controller that initially doesn't get an active
or standby assignment) will start up and register with AMF, but other than that
they will not do anything and just wait for a role change. When a role change
happens, they will continue with the rest of the start-up sequence and initalize
MDS so that they become visible on the network.

Locking of an OpenSAF 2N SU is currently not supported on a system configured
with more than two OpenSAF 2N SUs. This means that a controller node can start
up as a spare, but once it has received an active or standby assignment it
cannot go back to the spare role without a node reboot.