nid: Increase default time-out for clmna on system controllers [#1750]
Since standby and spare system controllers will hang in clmna until the active
CLM server accepts them into the cluster, the default clmna time-out has to be
set high enough to allow the CLM server on the active system controller to start
and IMM to finish loading. The default time-out for clmna is therefore increased
to the same as the time-out for clmd: 192000.