ntfa: return ERR_UNAVAILABLE on non-member node after headless state [#1744]
During headless state, OpenSAF may get stopped on payload with NTF app running.
Since OpenSAF is not running on the payload, any A.01.02 NTF client should not be served on
this node and this client should not be recovered. After first controller comes up, A.01.02
client will not be recovered and application will get SA_AIS_ERR_UNAVAILABLE upon which an
app can call saNtfFinalize() for freeing the resources.