amfnd: mark SU RESTARTING in comp FSM during restart of comps [#1752] V2
In reported problem, AMFD does not send state change notification for
1)SU presence state change from INSTANTIATED to RESTARTING and
2)SU presence state change from RESTARTING to INSTANTIATED
when component restarts due to RESTART admin op on it or faults with comp-restart recovery.

As per AMF spec, presence state of SU will be RESTARTING when all of its comp are in RESTARTING
presence state. In this case when comp restarts due to fault or RESTART admin op on it, AMFND
is not marking SU's presence state RESTARTING and SU remains in INSTANTIATED state. Since there
is no change in presence state of SU, no state change notification for SU is sent.

Patch fixes the problem by marking SU's presence state to RESTARTING when:
1)SU consists of a single restartable component and this comp restarts due to fault or RESTART
admin op on it.
2)SU consists of all restartable components and all these components faults simultaneously.