log: verify logBufSize to avoid node malfunctioned [#1789]
When accidentally passing an invalid value of logBufSize to saLogWriteLogAsync()
such as a very large number which is caused by not using strlen() on logBuf,
it will cause a lot of troubles.

Add code to verify if logBufSize is calculated based on logBuf or not
and also prevent too big data sent to server side.