smfd: include elapsedtime correctly in waitForNodeDestination and getNodeDestination [#1811]
smfd: include elapsedtime correctly in waitForNodeDestination and
getNodeDestination [#1811] The elapsed time is not correctly added to the
total time spent in the loops of the functions waitForNodeDestination and
getNodeDestination. This function ensures that the elapsed time is accounted
for correctly with relation to the rebootTimeout as configured by the user.
The patch also fixes a 20s deviation from the user configured rebootTimeout

Testing Commands:
-----------------
Configure a 60 second wait timeout in smfRebootTimeout attribute.
As a part of reboot, observe that that there is no deviation of 22 second in the
case of timeout before smfd declares timeout in waitinf for the rebooted node to comeup.