amfd: stop calling saImmOmInitialize in a loop on SU instantiation [#1819]
If immnd is down for an extended period of time, amfnd can be killed by the AMF watchdog
as it's stuck in immutil_saImmOmInitialize(). This can happen if a non OpenSAF component
restarts during this period.

saImmOmInitialize already has a potential timeout of 30s, and if we call saImmOmInitialize()
in a loop as we do in immutil_saImmOmInitialize() then a watchdog timeout is likely to be
triggered.