amf: start clm tracking after reading IMM config [#1815]
It has been observed that CLM callbacks to amfd can become 'lost'
in a large cluster. It seems to be occurring in MDS, when the callbacks are
sent around the same time as amfd is calling avd_imm_config_get().
A workaround is to call avd_clm_track_start() after avd_imm_config_get() is completed.
Further investigations are required. It seems avd_imm_config_get() generates a large
amount of traffic through MDS.