imm: Discard client connection when receiving A2ND_CL_TIMEOUT [#1817]
Add new member 'mPurged' to struct 'CcbInfo'.
It will be set to true when the ccb continuation is purged.
In the "ccb continuation" functions, mPurged is checked to
decide whether to response to client or not.

The value of 'mOriginatingConn' is needed to
get the CcbId when discarding the client (getCcbIdsForOrigCon).
So it will not be cleared when the ccb continuation is purged.