amfnd: fix COMP-FO recovery when cleanup time is more than sufailoverprob[#1839]
In the reported problem, fault of comp leads to comp-failover escalation.But
failover did not happen.

As a part of compt-failover, amfnd launched cleanup if component. Before clean up gets
completed, su-failover time expires and AMND resets escalation parameters. In this way
AMFND loses that the context of cleanup is comp-failver recovery. When comp gets cleaned up
successfully, AMFND (not ware of context) does inform AMFD for failover of assignments.

Patch fixes problem by remembering the comp or su failover context using a separate flag
and not relying on escalation params.