amfd: fix assignment of standby HA state without active HA state, NPM model [#1562]
When OpenSAF is stopped on one node hosting two application SUs, standby assignments
are given for SIs active on the stopped node.

In the reported problem, stopped node consists of active and standby SUs for two SIs.
When this node is stopped, AMFD tries to failover the SUs. For the active SU it deletes
the SUSI for both SIs. Since failover of this SU is not possible as standby assignment
also resides (susi to be deleted) on the stopped node only, AMFD runs new assignment logic.
Since no SU is available for active assignment for any SI, AMFD tries for fresh standby assignments.
In the logic of assigning standby HA state, AMFD checks if active assignments are there by checking
list_of_susi and also check if standby assignment is present. Here it must be noted that AMFD
checks only list_of_susi and not the HA state and it assumes that since list_of_susi is not null
it means active assignment are there. In the reported problem since both active and standby
assignments resides on the stopped node, AMFD gets list_of_susi non-null (contains susi for standby
HA state which will be deleted in failover logic of node). Since only one susi(assuming it active)
is present AMFD goes for standby assignments.

Patch ensures that before assignning fresh standby HA state, valid active assignment for
the SI must be verified because new assignment logic is also invoked during recovery phase.