plm: let hsm init plmc at system init [#1813] EE on which active PLMS is running can be shutdown when it receives its own plmc connection.
The plmc socket is initialized before HPI discovery has completed. In cases where HPI discovery takes a long time (with large numbers of FRUs), it is possible that the plmc connection from the local plmc can come in before this PLM has processed its own HE information from HPI. In this case PLMS will shut itself down because the HE is in OOS.

Let hsm init plmc if this is not a switchover.

Testing Commands:
-----------------
0) could be difficult to simulate (we see this in a live system)
1) you need an HPI inventory which is large (ours is 75 HEs)
2) start openhpid
3) start plmcd
4) start opensafd
5) pause HPI discovery in hsm for 5 minutes
6) let the system continue

Testing, Expected Results:
--------------------------
1) active PLMS should not shut itself down