amfd: do not send duplicate removal of assignment, 2N model [#1772]
In the reported problem, AMFND asserted when SU was unlocked.

For complete analysis, please refer ticket. In short, when AMFND was removing
the assignments, it gets a duplicate removal of assignment for the same SU because
of reboot of node hosting the active su. This duplicate message gets buffered and is picked
up when ongoing removal completes. After completion of ongoing removal of assignment, AMFND picks
buffered assignment and sets assignment related flags. Since SUSIs were deleted during previos
removal, no callbacks processing and response to AMFD is done for it. During response to AMFD,
AMFND resets all assignment related flags and it remained undone for buffered assignments.
Later on when SU was unlocked and fresh assignments were given to it. After completion of callback
when AMFND tries to respond to AMFND expects valid SI pointer for fresh assignment and checks it through
a assert statement. Here AMFND asserts because of side effects of assignment related flags being set.

Patch fixes the problem by avoiding sending duplicate removal of assignments to AMFND.