imm: Remove the poll timeout in IMM testcases [#1970]
Many IMM test-cases rely on poll timeout to end the while loop.
That will slow the test because it has to wait for the timeout.
Also the test-cases will fail if a callback doesn't come before the timeout occurs.
This patch removes those timeout and uses a pair of piped file-descriptors to stop the thread.