amfnd: send recovery request to amfd for term-failed su [#2003]
SU moved to term-failed state and repair admin opreration does not work.

When SG is unlocked, AMFND initiates active assignments by instantiating the
only component. After instantiation failure, AMFND tries to clean up the component.
Cleanup fails. AMFND marks comp and SU in TERM_FAILED state, but it neither responds
to AMFD for the completion of assignment nor it sends any recovery request.
Because of this SG remains unstable in REALIGN state.In this state,
no admin operation is allowed.

Patch solves the problem to send a recovery request to AMFD, so that it
deletes the assignments and marks SG stable.