log: fix ER syslog when running logtest 5 2 [#1864]
Not wait for apply callback in logtest 5 2 done before removing test directory
(in this test case, it changes 'logRootDirectory' attribute, then there are
action of renames files)

The patch changes to keep the test directory. Before creating test directory, if
test directory exists, removing it to make sure that new test directory is
created and cleaned.