amfd: do not send duplicate removal of assignment, N-Way model [#2022]
AMFD asserted during lock operation on NG when its nodes are deplyed with
N-WAY model application.

When lock on NG operation is issued, AMFD sends quiesced state to active SIs in all the
three SUs. Responses for SU2 and SU3 comes and AMFD sends SU level deletion to them.
SU2 contains standby assignment of SI1 which is still getting quiesced in SU1. Now
SU1 faults with su-failover recovery when SU2 is still not responded for deletion of
assignments. AMFD tries to perform recovery of SU1 and tries to fail-over SI1 of SU1
but could not find any valid standby in any other SU. Since failover is not
possible, AMFD tries to delete all standby assignments of SI1. AMFD tries to send
deletion for SI1 to SU2 for which SU level deletion is already sent and it asserts.

Patch avoids sending duplicate deletion of assignments.