imm: Do not send ccb abort reply to clients if they are not in CCB call [#2010]
If the ccb is in CCB_READY/CCB_VALIDATED/CCB_COMMITTED/CCB_ABORTED,
the client is not invoking any ccb call.
We do not send ccb abort reply to clients in that case.

In case of timeout, it's already guarded by 'mPurged'.

If ccb->mAugCcbParent is not null, the client must be in a CCB call,
so we don't have to check for ccb state in that case.
We always add 'ccb->mAugCcbParent->mOriginatingConn' in this case.