amfd: do not flush imm jobs on active controller[#2016]
AMF pushes IMM updates in a queue to be updated after finishing main job.
currently, standby amfd as a applier flushes job queue if number of elements
reaches to 200. Since check is on implementer status, active controller may
flush it when it loses implementer status and goes for reinit with IMM(one case
is when IMM returns BAD_HANDLE, a rare case).

Patch checks the role of amfd besides implemeter status before flusing the
job queue.