ntfs: convert LOG_ER to LOG_WA [#1895]
When unsubscribe() and finalize() do in parallel, ntfd may process finalize request first.
It delete client and anything relate to this client. When ntfd processes unsubscribe request,
It can't find clientId in the list  and the error is printed out here.

The patch makes to convert LOG_ER to LOG_WA.