log: fix runtime log stream is deleted after restarting from headless [#2071]
Did not look up in recovery list for openning streams, then there was case
the openning log stream did not exist in stream database but existing in
recovery list. So, creating this log stream would lead to ERR_EXIST (14)
and logsv deleted the stream from IMM.

This patch fixes this. Looking up in the recovery first, if not found
then creating log stream.