smf: balanced upgrade, modifications missing after reboot [#2147]
The issue happens duing balanced upgrade when doing an SI SWAP on the
controllers just before starting the balanced procedures. In this case SMF
normally would read the IMM data for any step modifications ticket [#906]
changed that to not read any modifications when the procedures are in COMPLETE
state. Solution is to keep the behaviour in a non balanced upgrade. When
executing a balanced upgrade the modifications are read and a error check
(forAddRemove and forModify) is skipped.