AMFND: Extend escalation support during headless at node director [#1902]
If any escalation requires a failover/switchover during headless, amfnd
currently reboot the node. This impacts on other healthy application
regardless reboot configuration. For example, if saAmfNodeAutoRepair is
disabled, that means there will not be a reboot if nodeFailover/Switchover

The patch removes the current reboot after failover/switchover is escalated
as of now. The purpose of the patch is to keep AMFND's behaviour staying
close to non-headless as much as possible. However, if componentFailover
of NPI SU is reached, it would be a suFailover orchestrated by AFMD.
But AMFD is not present, AMFND has to perform cleanup part as of suFailover
due to faulty component.