cpsv: on cpnd down fist remove child safReplica object then parent safCkpt object [#2189] Bug : While cpd processing cpnd down for  COLLOCATED  cktp  and that checkpoint only exist on the went down cpnd ( no others Node opened this checkpoint in cluster) , then cpd  removes  that checkpoint and replica completely.
In such case the current logic has as bug,  fist it removes ckpt node and
then replica, this is causing deletion of parent object safCkpt=...,*  first
, then child object safReplica=...,safCkpt=...,* next.

as we know IMM removes child if parent is removed ,so this is causing the
issue out of sequence remove of safReplica object and ERR_NOT_EXIST  is
returned.

Fix :
While cpd removing  that checkpoint and replica completely , follow the
sequence of  child object safReplica=...,safCkpt=...,*  fist then  parent
object safCkpt=...,* next.

This is focused fix , my be we need to review complete code for such
occurrences , if found will be addressed in new ticket.