base: Add Mutex, Lock and ConditionVariable classes [#2173]
Add classes for Mutex, Lock and ConditionVariable - similar to std::mutex,
std::unique_lock and std::condition_variable. One reason for adding these to
OpenSAF is that the C++ standard library casses are unapproved by the Google C++
Style Guide. Another reason is that we can integrate them better into OpenSAF:
we select the appropriate options when creating the mutex and condition
variable, and we call osaf_abort() in case an unexpected error is
encountered. Also, we will use an error checking mutex when the ENABLE_DEBUG
preprocessor macro is defined - paving the way for special OpenSAF debug builds.