amfd: increase max job queue size at standby [#2228]
After a cold sync, the standby job queue size can exceed the current limit of 200.
This results in the emptying of the entire queue at standby.

Potentially, some IMM updates are not done, as all jobs will be deleted during
a failover if the size is exceeded. This is especially important for updates that
the old 'active' could not perform in time during the failover.

This patch attempts to reduce the occurrence of this by increasing the max size,
and only trimming the queue to half the maximum size. The updates that must be done
are likely to be located towards the end of the queue, and these are kept.