amf: change saAmfCompCmdEnv attribute to writable [#2255]
Current OpenSAF implementation defines the attribute saAmfCompCmdEnv
in the SaAmfComp class as non-writable according to last AMF specification.
This restriction doesn't allow the upgrade of environment attributes defined
at component instance, or to remove it.

The attribute was made writable in an errata to the AMF model
(www.saforum.org/HOA/assn16627/images/SAI-IM-XMI-A.04.02.errata.xml.zip).
OpenSAF should comply to this change in the errata to enable the upgrade of
environment attributes defined at component instance level.