ntf: workaround to fix failed upgrade [#2300]
During cluster upgrade from an old OpenSAF version which have not included the
fix #2006 to one that have the fix in, if there is an code fault related to
inconsistent in additionalText and lengthAdditionalText in NTF application, could lead to upgadate failed.

This patch fix to add an workwround in NTFs, adjust lengthAdditionalText to correct one if NTFs find the mismatch

Note: The patch applies on branch 5.0 has little differences. There are some changes in test cases
ntftest 30 3... to change the length of additionalText+1  and lengthAdditionalText are the same.