build: Rename libsmf_common back to libsmfsv_common [#2298]
An application binary linked with the OpenSAF AIS libraries may fail to load
after upgrading to OpenSAF 5.2. The reason is that opensaf/libsmfsv_common.so.0
has been renamed to opensaf/libsmf_common.so.0. This problem can happen if the
application was linked without using the -Wl,--as-needed option (and this option
is not enabled by default by the Linux distribution used when building the
application binary).

Fix this by renaming opensaf/libsmf_common.so.0 back to
opensaf/libsmfsv_common.so.0. The documentation has already been updated by a
previous commit to mention that -Wl,--as-needed must to be used when linking
with the OpenSAF libraries, to avoid similar problems in the future.