imm: Fix problems with removing coordinator role when cluster goes headless [#2296]
When SC comes back too fast, it will fail to change to SEVER_READY state because immnd_proc_server() is not executed.
This patch basically reverts the changes in immnd_proc_server() made by #1692
and moves them to immnd_evt_proc_mds_evt().

Also, this patch hanldes the case when D2ND_SYNC_START is received but sync process isn't forked yet.
That case was not handled in the patch for #1692.