msgd: only call saClmFinalize if CLM init fails [#2308]
If a node is hosting a msg queue, and the node is rebooted, it can take 20
seconds for the msg queue to fail to another node.

saClmFinalize is mistakenly being called in clm_init, so CLM tracking
messages are never received for nodes that leave the cluster.

Only call saClmFinalize if CLM init fails.