base: fix ncs_edp_sanamet for arrays [#2314]
PLM and MSG standby do not come up in 5.1.0.

Both PLM and MSG make use of arrays of SaNameT structures. These are not
encoded and decoded properly by the new implementation of ncs_edp_sanamet.

A temporary solution is to put back the old functionality only for use
with arrays. Once it is clearly understood how the new function should
handle arrays, this fix can be removed.