logd: implement cfg files rotation [#2239] Requirement :
Current implementation of LOGsv do rotation of log files as defined, but not
for cfg files.
So, if we play with log streams application in a long time  a large numbers of
cfg files are generated ,
which is causing  system space issue.

Implementation :

SAI-AIS-LOG-A.02.01 specification not defied the approach on .cfg ROTATE
policy,
so it is implementation specific , this enhancement taken approach as CFG
file can also be ROTATED concurrently
whenever the log files with the same <closetime> suffix  is being ROTATED.

This enhancement ensures Until the old log files is not deleted and the old
.cfg file is still available since it contains information about how
to read the old log files, in other words when the last old log file has been
rotated out also the old .cfg
file will  be removed (but not before that).