amfd: Do not adjust saAmfSUAdminState [#2210]
If loss of update saAmfSUAdminState before headless occurs, amfd currently
does not have enough information to adjust saAmfSUAdminState. The reason is
that amfd can not differentiate whether si-swap or su-lock has been executed
before headless if saAmfSUAdminState was lost to update in IMM. Both si-swap
and su-lock are using the same SG FSM STATE, having QUIESCED assignment.

This patch prioritizes si-swap and ignore to update saAmfSUAdminState to LOCKED
in order to avoid out of service.