uml: Fix busybox build problem and upgrade Linux version [#2348]
UML failed to build on Fedora 24:

~~~
Trying libraries: crypt m
 Library crypt is not needed, excluding it
 Library m is not needed, excluding it
Final link with: <none>
Warning! Space in _end[] is too small (56 bytes)!
Rerun make to build a binary which doesn't use it!
Makefile:717: recipe for target 'busybox_unstripped' failed
~~~

This happened due to a tuning configuration option that ought to be
disabled. Fix this and also upgrade the Linux version to the latest:

* Linux 4.10.1
* Busybox 1.26.2
* Iproute2 4.10.0