smfd: fix campaign suspend when last step of procedure is executing [#2332]
Campaign is in suspended state (after ADMIN_SUSPEND sent to it), but no
procedure is in suspended state.

The code in SmfProcStateExecuting::executeStep and
SmfProcStateRollingBack::rollbackStep does not properly handle the last
step. If the last step completes the EXECUTE message is not sent back to the
procedure thread to see if a pending suspend is sitting there, but continues
on with the wrapup and returns COMPLETED. Then the waiting SUSPEND message
is executed, but the procedure is in COMPLETED state, so it is ignored.

Solution is to remove the code which checks if this is the last step.