base: Use dlopen to avoid LSB violation [#2396]
Ticket [#2266] introduced a dependency towards the openssl library. This causes
two problems: the first one is that this library is not part of LSB. The second
(and related) one is that OpenSAF binaries built on one Linux distribution may
fail to load on another Linux distribution, due to different names (versions) of
this library.

The solution uses dlopen() to open the library. If the library cannot be opened,
a simple fallback behaviour is used (the hash function returns all zeroes).