base: Use getdomainname() to build the FQDN [#2405]
Querying the DNS can potentially be very slow, especially in the case when the
DNS servers are not responding e.g. due to a misconfiguration. Therefore, we
instead use gethostname() + getdomainname() to build the fully qualified domain
name (FQDN) using locally stored information.

The result from getdomainname() may not always be accurate so the long-term
solution is probably to introduce a separate process/thread that can query the
DNS in the background.