smfd: fix race condition when detecting async failures [#2413]
smfd core dumps during commit of campaign.

If an AMF SU under maintenance fails right as the campaign commit is done, there
is a race condition present. Before SMF clears the suMaintenaceCampaign
attribute of the SU, if the SU fails, it will send a notification. Meanwhile,
the commit deletes upgrade campaign pointer inside smfd. If the deletion of the
campaign pointer inside smfd occurs before it receives the NTF event a crash
will occur because the campaign pointer is gone.

Solution is to always process NTF events before processing the termination of
the campaign. The campaign termination code sets "m_running" to false, and
deletes the pointer. This should always be last in the poll loop so that the
loop will exit immediately without processing any NTF events (or other future
events.)