smf: Attribute value handling in longDn applier is incorrect [#2442]
SMF has an IMM applier to track changes of attribute longDnsAllowed in OpensafImm class.
Fix that the applier does not find the value if the list of attributes to be handled in the
apply callback contains more than one attribute. Also if the found attribute is not the serached
one the cached attribute value is set to invalid